% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{add_colunas_por_extenso}
\alias{add_colunas_por_extenso}
\title{Adiciona colunas com nomes por extenso
respectivos aos destinos, situação, subsistemas e componentes}
\usage{
add_colunas_por_extenso(tabela)
}
\arguments{
\item{tabela}{tabela de dados com as necessidades de investimentos
com os seguintes campos: "destino";"situacao";"subsistema";"componente"}
}
\value{
tabela com as colunas adicionais "Destino", "Situação", "Subsistema" e "Componente"
}
\description{
Adiciona colunas com nomes por extenso
respectivos aos destinos, situação, subsistemas e componentes
}
