% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{add_density}
\alias{add_density}
\title{Calcula as densidades}
\usage{
add_density(df)
}
\arguments{
\item{data}{um tibble() contendo os dados do SNIS}
}
\value{
um tibble() contendo as densidades
}
\description{
Calcula as densidades
}
\examples{
\dontrun{
df <- add_density(df)
}
}
