% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{adiciona_atendimento_censo_2022}
\alias{adiciona_atendimento_censo_2022}
\title{Adiciona dados de atendimento do censo IBGE 2022}
\usage{
adiciona_atendimento_censo_2022(tabela, componente)
}
\arguments{
\item{tabela}{é um \code{data.frame} com no minimo a coluna codigo_municipio}

\item{componente}{é um \code{character} com o componente desejado (agua, esgoto ou residuos)}
}
\description{
Adiciona dados de atendimento do censo IBGE 2022
}
