% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pnad.R
\name{adiciona_atendimento_pnadc}
\alias{adiciona_atendimento_pnadc}
\title{Adiciona atendimento PNAD continua}
\usage{
adiciona_atendimento_pnadc(tabela, componente, ano)
}
\arguments{
\item{tabela}{contendo as regiao_sigla, populacao_total_corrente, populacao_urbana_corrente, populacao_rural_corrente}

\item{componente}{componente a ser adicionado}

\item{ano}{ano do pnad}
}
\value{
tabela com os dados de atendimento
}
\description{
Adiciona atendimento PNAD continua
}
