% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pnad.R
\name{adiciona_atendimento_relativo_pnadc}
\alias{adiciona_atendimento_relativo_pnadc}
\title{Adiciona atendimento relativo PNAD continua}
\usage{
adiciona_atendimento_relativo_pnadc(tabela, componente, ano)
}
\arguments{
\item{tabela}{contendo as regiao_sigla}

\item{componente}{componente a ser adicionado}

\item{ano}{ano do pnad}
}
\value{
tabela com os dados de atendimento
}
\description{
Adiciona atendimento relativo PNAD continua
}
