% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localidades.R
\name{adiciona_classificacao_litoranea}
\alias{adiciona_classificacao_litoranea}
\title{Classificação litorânea}
\usage{
adiciona_classificacao_litoranea(tabela)
}
\arguments{
\item{tabela}{contendo coluna codigo_municipio}
}
\value{
tabela com coluna adicional \code{litoral}
}
\description{
Adiciona se o municipio é litorâneo ou não.
}
\examples{
codigo_municipio <- c("1200013", "1200054", "1200104", "1200138")
input <- dplyr::tibble(codigo_municipio)
tabela <- adiciona_classificacao_litoranea(input)
}
