% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{adiciona_deficit_rural_agua_pnad}
\alias{adiciona_deficit_rural_agua_pnad}
\title{Déficit rural água PNAD}
\usage{
adiciona_deficit_rural_agua_pnad(tabela, deficit)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna \code{estado}}

\item{deficit}{um \code{data.frame} contendo a coluna \code{estado} e \code{deficit_agua_relativo_rural}}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{taxa_crescimento}.
}
\description{
Adiciona o déficit rural para sistema de abastecimento de água (PNAD).
}
