% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localidades.R
\name{adiciona_estado}
\alias{adiciona_estado}
\title{Adiciona coluna de estados na tabela}
\usage{
adiciona_estado(tabela)
}
\arguments{
\item{tabela}{contendo coluna codigo_municipio}
}
\value{
tabela com coluna adiciona estado
}
\description{
Adiciona coluna de estados na tabela
}
\examples{
codigo_municipio <- c("1200013", "1200054", "1200104", "1200138")
input <- dplyr::tibble(codigo_municipio)
tabela <- adiciona_estado(input)

}
