% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{adiciona_indices_drenagem}
\alias{adiciona_indices_drenagem}
\title{Índices de drenagem}
\usage{
adiciona_indices_drenagem(tabela)
}
\arguments{
\item{tabela}{contendo a coluna código do município}
}
\value{
a tabela de entrada com os campos adicionais dos índices (\code{caracteriticas_fisicas} e \code{infraestrutura})
}
\description{
Adiciona a tabela os índices de drenagem relativos as caracteristicas físicas e infraestrutura.
}
