% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localidades.R
\name{adiciona_pais}
\alias{adiciona_pais}
\title{Adiciona coluna de País}
\usage{
adiciona_pais(tabela)
}
\arguments{
\item{tabela}{}
}
\value{
tabela com uma coluna contendo a palabra "Brasil" em todas linhas
}
\description{
Adiciona na tabela de entrada uma coluna contendo o nome do País.
}
\examples{
codigo_municipio <- c("1200013", "1200054", "1200104", "1200138")
input <- dplyr::tibble(codigo_municipio)
tabela <- adiciona_pais(input)

}
