% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{adiciona_populacao_corrente}
\alias{adiciona_populacao_corrente}
\title{Junta as estimativas de populacao urbana}
\usage{
adiciona_populacao_corrente(populacao, ano, tabela)
}
\arguments{
\item{populacao}{é os dados da projeção populacional}

\item{ano}{da estimativa de população}

\item{tabela}{}
}
\value{
tabela com todos os dados consolidados
}
\description{
Junta as estimativas de populacao urbana
}
\examples{
\dontrun{
df <- adiciona_projecao_populacao(populacao, ano, tabela)
}
}
