% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{adiciona_preco_unidade_residuos}
\alias{adiciona_preco_unidade_residuos}
\title{Title}
\usage{
adiciona_preco_unidade_residuos(input, name, precos)
}
\arguments{
\item{name}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
