% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{adiciona_projecao_populacao}
\alias{adiciona_projecao_populacao}
\title{Junta as estimativas de populacao total, urbana, rural e dados de snis}
\usage{
adiciona_projecao_populacao(populacao, ano, tabela)
}
\arguments{
\item{populacao}{é os dados da projeção populacional}

\item{ano}{da estimativa de população}

\item{snis}{dados do snis}
}
\value{
tabela com todos os dados consolidados
}
\description{
Junta as estimativas de populacao total, urbana, rural e dados de snis
}
\examples{
\dontrun{
df <- adiciona_projecao_populacao(populacao, ano, tabela)
}
}
