% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localidades.R
\name{adiciona_regiao}
\alias{adiciona_regiao}
\title{Adiciona coluna de regiões}
\usage{
adiciona_regiao(tabela)
}
\arguments{
\item{tabela}{contendo coluna codigo_municipio}
}
\value{
tabela com coluna adicional regiao
}
\description{
Adiciona a tabela de entrada um coluna contendo as sigla da região em que o município se encontra.
}
\examples{
codigo_municipio <- c("1200013", "1200054", "1200104", "1200138")
input <- dplyr::tibble(codigo_municipio)
tabela <- adiciona_regiao(input)

}
