% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{adiciona_seguranca_hidrica}
\alias{adiciona_seguranca_hidrica}
\title{Segurança Hídrica}
\usage{
adiciona_seguranca_hidrica(tabela, seguranca_hidrica)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna \code{codigo_municipio}}

\item{deficit}{um \code{data.frame} contendo a coluna \code{codigo_municipio}, \code{seguranca_hidrica} e \code{semiarido}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{taxa_crescimento}.
}
\description{
Adiciona dados de segurança hídrica
}
