% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{adiciona_taxa_crescimento}
\alias{adiciona_taxa_crescimento}
\title{Taxa de crescimento populacional}
\usage{
adiciona_taxa_crescimento(tabela, projecao)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna codigo_municipio}

\item{projecao}{um \code{data.frame} contendo a coluna \code{codigo_municipio} e \code{taxa_de_crescimento}}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{taxa_crescimento}.
}
\description{
Adiciona taxa de crescimento populacional
}
