% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{adiciona_tipo_disposicao}
\alias{adiciona_tipo_disposicao}
\title{Adiciona tipo de disposição de resíduos}
\usage{
adiciona_tipo_disposicao(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} com a coluna \code{codigo_municipio}}
}
\value{
um \code{data.frame} com as colunas adicionais \code{meta_coleta}, \code{meta_compostagem}e \code{meta_reaproveitamento}
}
\description{
Adiciona tipo de disposição de resíduos
}
