% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{agua_preprocess}
\alias{agua_preprocess}
\title{Agua Pre-processing
Read the SINISA water data and process it into a data frame.}
\usage{
agua_preprocess(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\value{
A data frame containing the processed water data
}
\description{
Agua Pre-processing
Read the SINISA water data and process it into a data frame.
}
