% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\docType{data}
\name{agua_required_fields}
\alias{agua_required_fields}
\title{Necessidade de Investimento - Componente Água - Situação Urbana}
\format{
An object of class \code{character} of length 10.
}
\usage{
agua_required_fields
}
\description{
Necessidade de Investimento - Componente Água - Situação Urbana
}
\keyword{datasets}
