% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{aguas_pluviais_preprocess}
\alias{aguas_pluviais_preprocess}
\title{Aguas Pluviais Pre-processing
Read the SINISA rainwater data and process it into a data frame.}
\usage{
aguas_pluviais_preprocess(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\value{
A data frame containing the processed rainwater data
}
\description{
Aguas Pluviais Pre-processing
Read the SINISA rainwater data and process it into a data frame.
}
