% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{anos_censo_base_calculo}
\alias{anos_censo_base_calculo}
\title{Retorna os anos dos censos disponíveis na base de cálculo}
\usage{
anos_censo_base_calculo()
}
\value{
Um vetor com os anos dos censos disponíveis.
}
\description{
Esta função lista os arquivos na pasta 'dados/base_calculo' que contêm o padrão "censo" no nome
e extrai os anos correspondentes.
}
