% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pnad.R
\name{anos_disponiveis_pnadc}
\alias{anos_disponiveis_pnadc}
\title{Extrai os anos disponiveis na PNADC para um componente}
\usage{
anos_disponiveis_pnadc(componente)
}
\arguments{
\item{componente}{Componente a ser verificado ("agua", "esgoto", residuos")}
}
\value{
Vetor de anos disponíveis
}
\description{
Extrai os anos disponiveis na PNADC para um componente
}
