% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{anos_do_censo_disponiveis_para_download}
\alias{anos_do_censo_disponiveis_para_download}
\title{Lista de anos da população estimada pelo Censo
que estão disponíveis para download no ftp do IBGE}
\usage{
anos_do_censo_disponiveis_para_download()
}
\value{
lista de anos do censo disponíveis
}
\description{
Lista de anos da população estimada pelo Censo
que estão disponíveis para download no ftp do IBGE
}
\examples{
anos_do_censo_disponiveis_para_download
}
