% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{aplica_regressao_drenagem}
\alias{aplica_regressao_drenagem}
\title{Aplica regressao simples drenagem}
\usage{
aplica_regressao_drenagem(tabela, modelo)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas GE006}

\item{modelo}{um objeto da classe \code{lm} contendo os parâmetros do modelo de regressão}
}
\value{
tabela com os valores de investimento
}
\description{
Aplica regressao simples drenagem
}
