% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{app_state_exists}
\alias{app_state_exists}
\title{Verifica se o arquivo de estado da aplicação ja foi salvo}
\usage{
app_state_exists()
}
\value{
\code{TRUE} se o arquivo existe e \code{FALSE}caso não exista.
}
\description{
Verifica se o arquivo de estado da aplicação ja foi salvo
}
\examples{
app_state_exists()
}
