% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{app_state_filename}
\alias{app_state_filename}
\title{Local do arquivo de estado da aplicação}
\usage{
app_state_filename()
}
\value{
o caminho do arquivo de estado da aplicação
}
\description{
Local do arquivo de estado da aplicação
}
\examples{
path <- app_state_file_name()
}
