% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_municipio}
\alias{area_municipio}
\title{Áreas territoriais dos municípios brasileiros}
\format{
Um data frame com 1 municipios por linha e 2 colunas, sendo estas:
\describe{
\item{codigo_municipio}{Código IBGE do município}
\item{area}{Área do município em km²}
}
}
\source{
\url{https://geoftp.ibge.gov.br/organizacao_do_territorio/estrutura_territorial/areas_territoriais/}
}
\usage{
area_municipio
}
\description{
Dataset contendo as áreas dos muncípios brasileiros em km²
}
\keyword{datasets}
