% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{area_urbana}
\alias{area_urbana}
\title{Área urbana}
\usage{
area_urbana(tabela)
}
\arguments{
\item{tabela}{contendo as colunas do SNIS \code{GE002}}
}
\value{
uma tabela contendo a densidade urbana
}
\description{
Área urbana
}
\examples{
\dontrun{
tabela <- area_urbana(tabela)
}
}
