% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_urbana_municipio}
\alias{area_urbana_municipio}
\title{Áreas urbanas dos municípios brasileiros}
\format{
Um data frame com 1 municipios por linha e 2 colunas, sendo estas:
\describe{
\item{codigo_municipio}{Código IBGE do município}
\item{area_urbana}{Área urbana do município em km²}
}
}
\source{
\url{https://biblioteca.ibge.gov.br/index.php/biblioteca-catalogo?view=detalhes&id=2100639}
}
\usage{
area_urbana_municipio
}
\description{
Dataset contendo as áreas urbanas dos muncípios brasileiros em km²
}
\keyword{datasets}
