% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{atendimento_relativo_residuos}
\alias{atendimento_relativo_residuos}
\title{Title}
\usage{
atendimento_relativo_residuos(tabela)
}
\arguments{
\item{tabela}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
