% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localidades.R
\name{base_municipios}
\alias{base_municipios}
\title{Cria tabela de localidades}
\usage{
base_municipios()
}
\value{
tabela com as localidades
}
\description{
Carrega a tabela de localidades do pacote rsan e retorna uma tabela com os campos
\code{codigo_municipio}, \code{codigo_localidade}, \code{localidade}, \code{estado} e \code{pais}.
}
