% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{calcula_capacidade_instalada}
\alias{calcula_capacidade_instalada}
\title{Title}
\usage{
calcula_capacidade_instalada(snis, custo)
}
\arguments{
\item{custo}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
