% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{calcula_custo_expansao}
\alias{calcula_custo_expansao}
\title{Calcula o custo de expansão do sistema de água e esgoto}
\usage{
calcula_custo_expansao(
  demanda,
  distribuicao,
  coleta,
  producao,
  tratamento,
  perda_agua
)
}
\arguments{
\item{demanda}{tabela com as demandas de distribuição e produção de água, coleta e tratamento de egoto.}

\item{distribuicao}{tabela com custos relativos (R$/m) de distribuição de água por estado}

\item{coleta}{tabela com custos relativos (R$/m) de coleta de esgoto por estado}

\item{producao}{tabela com custos relativos R$/(m3/ano) de produção de água}

\item{tratamento}{tabela com custos relativos R$/(m3/ano) de tratamento de esgoto}

\item{perda_agua}{Parâmetro de correção de perda de água em (\%)}
}
\value{
tabela com as necessidade de investimento
}
\description{
Calcula o custo de expansão do sistema de água e esgoto
}
\examples{
\dontrun{
invest <- calcula_custo_expansao(demanda, distribuicao, coleta, producao, tratamento, perda_agua)
}
}
