% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{calcula_custo_expansao_agua}
\alias{calcula_custo_expansao_agua}
\title{Calcula o custo de expansão do sistema de água}
\usage{
calcula_custo_expansao_agua(demanda, distribuicao, producao, perda_agua)
}
\arguments{
\item{demanda}{tabela com as demandas de distribuição e produção de água, coleta e tratamento de egoto.}

\item{distribuicao}{tabela com custos relativos (R$/m) de distribuição de água por estado}

\item{producao}{tabela com custos relativos R$/(m3/ano) de produção de água}

\item{perda_agua}{Parâmetro de correção de perda de água em (\%)}
}
\value{
tabela com as necessidade de investimento
}
\description{
Calcula o custo de expansão do sistema de água
}
\examples{
\dontrun{
invest <- calcula_custo_expansao_agua(demanda, distribuicao, coleta, producao, tratamento, perda_agua)
}
}
