% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{calcula_custo_expansao_esgoto}
\alias{calcula_custo_expansao_esgoto}
\title{Calcula o custo de expansão do sistema de esgoto}
\usage{
calcula_custo_expansao_esgoto(demanda, coleta, tratamento)
}
\arguments{
\item{demanda}{tabela com as demandas de distribuição e produção de água, coleta e tratamento de egoto.}

\item{coleta}{tabela com custos relativos (R$/m) de coleta de esgoto por estado}

\item{tratamento}{tabela com custos relativos R$/(m3/ano) de tratamento de esgoto}
}
\value{
tabela com as necessidade de investimento
}
\description{
Calcula o custo de expansão do sistema de esgoto
}
\examples{
\dontrun{
invest <- calcula_custo_expansao_esgoto(
  demanda, distribuicao, coleta, producao, tratamento, perda_agua
)
}
}
