% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{calcula_custo_expansao_producao}
\alias{calcula_custo_expansao_producao}
\title{Calcula a necessidade de investimento para produção de água}
\usage{
calcula_custo_expansao_producao(demanda, producao, perda_agua)
}
\arguments{
\item{demanda}{tabela com as demandas por município}

\item{producao}{tabela com os custos das unidades de produção de água}

\item{perda_agua}{perda de água a ser compensada (\%)}
}
\value{
tabela com a necessidade de investimento para produção de água
}
\description{
Calcula a necessidade de investimento para produção de água
}
\examples{
\dontrun{
custo <- calcula_custo_expansao_producao(demanda, producao, perda_agua)
}
}
