% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto.R
\name{calcula_custo_expansao_tratamento}
\alias{calcula_custo_expansao_tratamento}
\title{Calcula a necessidade de investimento para tratamendo de esgoto}
\usage{
calcula_custo_expansao_tratamento(demanda, tratamento)
}
\arguments{
\item{demanda}{tabela com as demandas de tratamento de esgoto (m3/ano) por município}

\item{tratamento}{tabela com os custos das unidades de tratamento de esgoto por município}
}
\value{
tabela com a  necessidade de investimento para tratamendo de esgoto
}
\description{
Calcula a necessidade de investimento para tratamendo de esgoto
}
\examples{
\dontrun{
custo_relativo <- calcula_custo_expansao_tratamento(demanda, tratamento)
}
}
