% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{calcula_custo_extensao}
\alias{calcula_custo_extensao}
\title{Calcula o custo de expansão para distribuição de água ou coleta de esgoto}
\usage{
calcula_custo_extensao(demanda, extensao, tipo)
}
\arguments{
\item{demanda}{tabela com a demanda de distribuição de água}

\item{extensao}{tabela com o custo relativo de distribuição de água}

\item{tipo}{define o tipo de custo (\code{"distribuicao_agua"} ou \code{"coleta_esgoto"}).}
}
\value{
o custo de expansão para o tipo escolhido
}
\description{
Calcula o custo de expansão para distribuição de água ou coleta de esgoto
}
\examples{
\dontrun{
custo <- calcula_custo_extensao(demanda, extensao, tipo)
}
}
