% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{calcula_custo_relativo_producao}
\alias{calcula_custo_relativo_producao}
\title{Calcula custo relativo para produção de água}
\usage{
calcula_custo_relativo_producao(preco_unidade, projeto_producao)
}
\arguments{
\item{preco_unidade}{tabela contendo as colunas estado, unidade e preco}

\item{projeto_tratamento}{tabela contendo cenario, unidade e quantidade}
}
\value{
tabela com as colunas estado, cenario e custo_relativo
}
\description{
Custo relativo é o valor monetário necessário para produzir 1m3 de água tratada por ano R$/(m3/ano).
}
\examples{
estado <- c("AC", "AC", "AC", "AL", "AL", "AL")
unidade <- c("ETA200", "EEA200", "POÇO40", "ETA200", "EEA200", "POÇO40")
preco <- c(3.055, 0.129, 0.69, 2.88, 0.12, 0.66)
preco_unidade <- dplyr::tibble(estado, unidade, preco)

cenario <- c("08", "08", "08", "08", "09", "09", "09", "09")
unidade <- c("ETA200", "EEA200", "POÇO40", "EEA200", "ETA200", "EEA200", "POÇO40", "EEA200")
tipo <- c("superficial", "superficial", "subterranea", "subterranea", "superficial", "superficial", "subterranea", "subterranea")
quantidade <- c(0.854106686284393, 1.3, 1, 1, 0.549559117416313, 1.3, 1, 1)
projeto_producao <- dplyr::tibble(cenario, unidade, tipo, quantidade)
df_out <- calcula_custo_relativo_producao(preco_unidade, projeto_producao)
}
