% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{calcula_deficit_agua_relativo_pnadc}
\alias{calcula_deficit_agua_relativo_pnadc}
\title{Calcula deficit relativo para agua situação rural usando dados da pnadc}
\usage{
calcula_deficit_agua_relativo_pnadc(tabela, ano)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{populacao} e \code{atendimento_rur_agua}.}

\item{ano}{um \code{number} contendo o ano do PNADC.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{deficit_agua_relativo_rural}.
}
\description{
Calcula deficit relativo para agua situação rural usando dados da pnadc
}
