% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{calcula_deficit_esgoto_relativo_censo}
\alias{calcula_deficit_esgoto_relativo_censo}
\title{Calcula deficit relativo para agua situação rural usando dados do censo}
\usage{
calcula_deficit_esgoto_relativo_censo(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{populacao} e \code{atendimento_esgoto}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{deficit_esgoto_relativo_rural}.
}
\description{
Calcula deficit relativo para agua situação rural usando dados do censo
}
