% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{calcula_deficit_esgoto_relativo_pnadc}
\alias{calcula_deficit_esgoto_relativo_pnadc}
\title{Calcula deficit relativo para agua situação rural usando dados do pnadc}
\usage{
calcula_deficit_esgoto_relativo_pnadc(tabela, ano)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{populacao} e \code{atendimento_esgoto_relativo_rural}.}

\item{ano}{um \code{integer} com o ano de referência para os dados do PNADC.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{deficit_esgoto_relativo_rural}.
}
\description{
Calcula deficit relativo para agua situação rural usando dados do pnadc
}
