% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{calcula_fator_correcao_sinapi}
\alias{calcula_fator_correcao_sinapi}
\title{Calcula o fator de correção monetária baseado no índice IGP}
\usage{
calcula_fator_correcao_sinapi(sinapi, ano_corrente)
}
\arguments{
\item{sinapi}{Uma string representando o nome do arquivo SINAPI, que inclui o ano e o mês no formato "SINAPI_YYYYMM".}

\item{ano_corrente}{Um número representando o ano corrente para o cálculo do fator de correção.}
}
\value{
Um valor numérico representando o fator de correção monetária.
}
\description{
Esta função calcula o fator de correção monetária para ajustar os preços de acordo com o índice IGP (Índice Geral de Preços).
}
\examples{
\dontrun{
fator <- calcula_fator_correcao_sinapi("SINAPI_202112", 2025)
}
}
