% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{calcula_preco_unidades_producao}
\alias{calcula_preco_unidades_producao}
\title{Calcula o preço das unidades de produção de água}
\usage{
calcula_preco_unidades_producao(
  df_projeto,
  sinapi,
  fator_insumo = 26,
  fator_composicao = 18,
  sub_total = TRUE
)
}
\arguments{
\item{df_projeto}{tabela com os dados do projeto tipo de produção de água}

\item{sinapi}{tabela com os dados do SINAPI}

\item{fator_insumo}{fator de correção do preços na categoria insumo (\%)}

\item{fator_composicao}{fator de correção do preços na categoria composição (\%)}

\item{sub_total}{Se FALSE a porcentagem é calculada como juros sobre juros. Se TRUE a porcentagem é a soma dos juros.}
}
\value{
uma tabela com os preços das unidades de produção de água por estado e cenário
}
\description{
Calcula o preço das unidades de produção de água
}
\examples{
\dontrun{
tabela <- calcula_preco_unidades_producao(df_projeto, sinapi, 26, 18, TRUE)
}
}
