% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{calcula_projecao}
\alias{calcula_projecao}
\title{Calcula a projeção populacional para cada ano dentro do período requerido.}
\usage{
calcula_projecao(tabela, ano_inicial, ano_final)
}
\arguments{
\item{tabela}{Um tibble() contendo os dados de populacao (total, rural, urbana) e taxa de crescimento}

\item{ano_inicial}{é o ano inicial da simulação (ano da fonte 2)}

\item{ano_final}{é o último ano que será simulado}
}
\value{
Um tibble() contendo a projeção populacional.
}
\description{
Calcula a projeção populacional para cada ano dentro do período requerido.
}
\examples{
\dontrun{
tabela <- calcula_projecao(tabela, 2021, 2033)
}
}
