% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{calcula_reposicao_total}
\alias{calcula_reposicao_total}
\title{Calcula reposicação total}
\usage{
calcula_reposicao_total(
  tabela,
  campo_capacidade,
  campo_investimento,
  campo_reposicao,
  ano_inicial,
  ano_final,
  vida_util
)
}
\arguments{
\item{tabela}{tabela contendo os dados de capacidade e investimento}

\item{campo_capacidade}{nome do campo de capacidade}

\item{campo_investimento}{nome do campo de investimento}

\item{campo_reposicao}{nome do campo de saida com o valor do custo de reposição total}

\item{ano_inicial}{ano da estimativa de capacidade instalada}

\item{ano_final}{último ano de projeção}

\item{vida_util}{vida útil média dos ativos (anos)}
}
\value{
tabela contendo a coluna adiciona de reposição
}
\description{
A reposição total refere-se ao investimento em reposição total entre o ano inicial e o ano final.
}
\examples{
investimento <- c(1,3)
capacidade <-   c(0,1)
df_in <- dplyr::tibble(capacidade, investimento)
df_out <- calcula_reposicao_total(df_in,"capacidade","investimento","reposicao", 2021, 2033, 30)
}
