% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{calcula_taxa_crescimento}
\alias{calcula_taxa_crescimento}
\title{Calcula a taxa de crescimento populacional
Utiliza método de crescimento geométrico}
\usage{
calcula_taxa_crescimento(tabela, ano_fonte1, ano_fonte2)
}
\arguments{
\item{tabela}{é um tibble() contentando os dados consolidados de população}

\item{ano_fonte1}{é o ano de obtenção da população (fonte 1)}

\item{ano_fonte2}{é o ano de obtenção da população (fonte 2)}
}
\value{
Um tibble() contendo a coluna taxa_de_crescimento
}
\description{
Calcula a taxa de crescimento populacional
Utiliza método de crescimento geométrico
}
\examples{
\dontrun{
df <- calcula_taxa_crescimento(df, 2010, 2021)
}
}
