% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{calcular_urbana_rural_fonte2}
\alias{calcular_urbana_rural_fonte2}
\title{Calcula a população urbana e rural da fonte 2
Assume-se que a porcentagem obtida na fonte 1 não é alterada
Assim necessita que a fonte 1 tenha os campos:
relativa_urbana e relativa_rural calculados pela funcao
adicionar_proporcao_urbana_rural()}
\usage{
calcular_urbana_rural_fonte2(data)
}
\arguments{
\item{data}{Um tibble() contendo os dados de populacao}
}
\value{
Um tibble() contendo as colunas populacao_urbana_fonte2 e populacao_rural_fonte2
}
\description{
Calcula a população urbana e rural da fonte 2
Assume-se que a porcentagem obtida na fonte 1 não é alterada
Assim necessita que a fonte 1 tenha os campos:
relativa_urbana e relativa_rural calculados pela funcao
adicionar_proporcao_urbana_rural()
}
\examples{
\dontrun{
df <- calcular_urbana_rural_fonte2(df)
}
}
