% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{calculate_demografico_agua}
\alias{calculate_demografico_agua}
\title{Calcula deficits e demandas para água}
\usage{
calculate_demografico_agua(df, meta_agua)
}
\arguments{
\item{df}{um \code{data.frame} contendo os dados do SNIS e de população}

\item{meta_agua}{meta em \% de atendimento para abastecimento de água}
}
\value{
um \code{data.frame} contendo os dados de deficits e demandas
}
\description{
Calcula deficits e demandas para água
}
\examples{
\dontrun{
df <- calculate_demografico_agua(df, 99, 90, 80)
}
}
