% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{calculate_demografico_esgoto}
\alias{calculate_demografico_esgoto}
\title{Calcula deficits e demandas para esgoto}
\usage{
calculate_demografico_esgoto(df, meta_esgoto, proporcao)
}
\arguments{
\item{df}{um \code{data.frame} contendo os dados do SNIS e de população}

\item{meta_esgoto}{meta em \% de atendimento para tratamento de esgoto}

\item{proporcao}{proporção da densidade de água que equivale a densiade de esgoto}
}
\value{
um \code{data.frame} contendo os dados de deficits e demandas
}
\description{
Calcula deficits e demandas para esgoto
}
\examples{
\dontrun{
df <- calculate_demografico_esgoto(df, 99, 90, 80)
}
}
