% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{calculate_geografico}
\alias{calculate_geografico}
\title{Calcula deficits e demandas para água e esgoto}
\usage{
calculate_geografico(df, meta_agua, meta_esgoto, proporcao)
}
\arguments{
\item{df}{um tibble() contendo os dados do SNIS e de população}

\item{meta_agua}{meta em \% de atendimento para abastecimento de água}

\item{meta_esgoto}{meta em \% de atendimento para tratamento de esgoto}

\item{proporcao}{proporção da densidade de água que equivale a densiade de esgoto}
}
\value{
um tibble() contendo os dados de deficits e demandas
}
\description{
Calcula deficits e demandas para água e esgoto
}
\examples{
\dontrun{
df <- calculate_geografico(df, 99, 90, 80)
}
}
