% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{caminho_do}
\alias{caminho_do}
\title{Retorna o caminho de um conjunto de dados}
\usage{
caminho_do(nome)
}
\arguments{
\item{nome}{é o nome do conjunto de dados}
}
\value{
o local do arquivo
}
\description{
Retorna o caminho de um conjunto de dados
}
