% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reposicao.R
\name{campos_reposicao}
\alias{campos_reposicao}
\title{Parâmetros de reposição}
\usage{
campos_reposicao(capacidade, custo, reposicao, vida_util)
}
\arguments{
\item{capacidade}{é um \code{character} com o nome dado para coluna capacidade instalada}

\item{custo}{é um \code{character} com o nome dado para coluna de custo unitário}

\item{reposicao}{é um \code{character} com o nome que será dado para coluna de reposição}

\item{vida_util}{é um \code{numeric} com o valor da vida util da instalação}
}
\value{
uma lista com todos parâmetros necessários para o cálculo de reposição
}
\description{
Retorna uma lista com os parâmetros para o cálculo da reposição
}
\examples{
params <- campos_reposicao("capacidade", "custo", "reposicao", 5)
}
