% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{capacidade_instalada_aterro}
\alias{capacidade_instalada_aterro}
\title{Capacidade instalada aterro}
\usage{
capacidade_instalada_aterro(tabela, vida_util)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas total_residuos, preco_unidade_aterro}

\item{vida_util}{um campo numérico contendo a vida util do aterro}
}
\value{
um \code{data.frame} contendo o valor da capacidade instalada do aterro (\code{capacidade_instalada_aterro})
}
\description{
Capacidade instalada aterro
}
\examples{
\dontrun{
tabela <- capacidade_instalada_aterro(tabela, 20)
}
}
