% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{capacidade_instalada_coleta_indiferenciada}
\alias{capacidade_instalada_coleta_indiferenciada}
\title{Title}
\usage{
capacidade_instalada_coleta_indiferenciada(tabela, valor)
}
\arguments{
\item{valor}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
