% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{capacidade_instalada_compostagem}
\alias{capacidade_instalada_compostagem}
\title{Capacidade instalada de compostagem}
\usage{
capacidade_instalada_compostagem(tabela, vida_util)
}
\arguments{
\item{vida_util}{}
}
\description{
Capacidade instalada de compostagem
}
\examples{
\dontrun{

}
}
