% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{capacidade_instalada_drenagem}
\alias{capacidade_instalada_drenagem}
\title{Capacidade instalada para drenagem}
\usage{
capacidade_instalada_drenagem(tabela)
}
\arguments{
\item{tabela}{\code{data.frame} contendo os campos:}
}
\value{
um \code{data.frame} contendo o campo \verb{capacidade instalada}
}
\description{
Capacidade instalada para drenagem
}
\examples{
\dontrun{
tabela <- capacidade_instalada_drenagem(tabela)
}
}
