% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto.R
\name{capacidade_instalada_esgoto}
\alias{capacidade_instalada_esgoto}
\title{Capacidade instalada sistema de abastecimento de água}
\usage{
capacidade_instalada_esgoto(snis, custo)
}
\arguments{
\item{snis}{um \code{data.frame} contendo as colunas \code{codigo_municipio}, \code{extensao_rede_esgoto_km} e \code{volume_esgoto_tratado_dam3_ano}}

\item{custo}{um \code{data.frame} contendo as colunas \code{codigo_municipio}, \code{custo_relativo_tratamento}, \code{preco_coleta_esgoto}}
}
\value{
um \code{data.frame} contendo as colunas \code{capacidade_instalada_coleta} e \code{capacidade_instalada_tratamento}
}
\description{
Capacidade instalada sistema de abastecimento de água
}
