% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{capacidade_instalada_rural_agua}
\alias{capacidade_instalada_rural_agua}
\title{Capacidade instalada Água Rural}
\usage{
capacidade_instalada_rural_agua(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas:
\code{domicilios_adequados_agua}, \code{custo_agua_distribuicao_rural} e \code{custo_agua_producao_rural}}
}
\value{
um \code{data.frame} contendo as colunas adicional capacidade_instalada_distribuicao_agua  e capacidade_instalada_producao_agua.
}
\description{
Calcula a capacidade instalada para expansão do sistema de abastecimento de água rural.
}
