% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{capacidade_instalada_rural_esgoto}
\alias{capacidade_instalada_rural_esgoto}
\title{Capacidade instalada esgoto situação rural}
\usage{
capacidade_instalada_rural_esgoto(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas:
\code{domicilios_adequados_esgoto}, \code{custo_esgoto_distribuicao_rural} e \code{custo_esgoto_producao_rural}}
}
\value{
um \code{data.frame} contendo as colunas adicional capacidade_instalada_coleta_esgoto  e capacidade_instalada_tratamento_esgoto.
}
\description{
Calcula a capacidade instalada para expansão do sistema de abastecimento de esgoto situação rural.
}
