% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{capacidade_instalada_triagem}
\alias{capacidade_instalada_triagem}
\title{Capacidade instalada de triagem}
\usage{
capacidade_instalada_triagem(tabela, vida_util)
}
\arguments{
\item{vida_util}{}
}
\description{
Capacidade instalada de triagem
}
\examples{
\dontrun{

}
}
