% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{carrega_base_calculo}
\alias{carrega_base_calculo}
\title{Carrega a base de cálculo da Necessidade de Investimento em Saneamento}
\usage{
carrega_base_calculo(componente, fonte, ano)
}
\arguments{
\item{componente}{Componente da base de cálculo}

\item{fonte}{Fonte dos dados}

\item{ano}{Ano da base de cálculo}
}
\value{
Data frame com os dados da base de cálculo
}
\description{
Carrega a base de cálculo da Necessidade de Investimento em Saneamento
}
