% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{carrega_censo}
\alias{carrega_censo}
\title{Carrega dados do censo IBGE 2022 agregado por municipio}
\usage{
carrega_censo(year)
}
\arguments{
\item{year}{ano que deseja carregar}
}
\value{
um \code{data.frame} com a estimativa populacional pelo Censo
}
\description{
Carrega dados do censo IBGE 2022 agregado por municipio
}
