% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{carrega_censo2022}
\alias{carrega_censo2022}
\title{Carrega dados do censo IBGE 2022 agregado por municipio}
\usage{
carrega_censo2022()
}
\value{
um \code{data.frame} com a estimativa populacional pelo Censo
}
\description{
Carrega dados do censo IBGE 2022 agregado por municipio
}
