% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{carrega_censo2022_setor}
\alias{carrega_censo2022_setor}
\title{Carrega dados censo IBGE 2022 agregado por setor}
\usage{
carrega_censo2022_setor()
}
\value{
um \code{data.frame} com a estimativa populacional pelo Censo por setor censitário
}
\description{
Carrega dados censo IBGE 2022 agregado por setor
}
