% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{carrega_censo_setor}
\alias{carrega_censo_setor}
\title{Carrega dados censo IBGE 2022 agregado por setor}
\usage{
carrega_censo_setor(year)
}
\arguments{
\item{year}{ano que deseja carregar}
}
\value{
um \code{data.frame} com a estimativa populacional pelo Censo por setor censitário
}
\description{
Carrega dados censo IBGE 2022 agregado por setor
}
