% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{carrega_cobertura_rede}
\alias{carrega_cobertura_rede}
\title{Carrega dados de cobertura de rede de água do censo IBGE 2022}
\usage{
carrega_cobertura_rede()
}
\value{
um \code{data.frame} com a cobertura de rede de água por município
}
\description{
Carrega dados de cobertura de rede de água do censo IBGE 2022
}
