% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{carrega_dado_auxiliar}
\alias{carrega_dado_auxiliar}
\title{Carrega um dado auxiliar da pasta dados/base_calculo}
\usage{
carrega_dado_auxiliar(nome)
}
\arguments{
\item{nome}{Nome do arquivo (sem extensão .csv) a ser carregado.}
}
\value{
Um data frame com os dados do arquivo auxiliar, ou NULL se o arquivo não for encontrado.
}
\description{
Esta função lê um arquivo CSV auxiliar localizado em 'dados/base_calculo' e retorna um data frame.
Caso a coluna 'codigo_municipio' exista, ela será convertida para caractere.
}
