% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{carrega_dado_populacao}
\alias{carrega_dado_populacao}
\title{Carrega dados populacionais de um arquivo auxiliar}
\usage{
carrega_dado_populacao(nome)
}
\arguments{
\item{nome}{Nome do arquivo (sem extensão .csv) a ser carregado.}
}
\value{
Um data frame com os dados populacionais filtrados, ou NULL se o arquivo não for encontrado.
}
\description{
Esta função lê um arquivo CSV de dados populacionais localizado em 'dados/base_calculo' e retorna um data frame
contendo apenas as colunas permitidas: codigo_municipio, municipio, populacao_rural, populacao_urbana, populacao_total.
Caso a coluna 'codigo_municipio' exista, ela será convertida para caractere.
}
