% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{carrega_dados_sinisa}
\alias{carrega_dados_sinisa}
\title{Carrega os dados pré-processados do SINISA}
\usage{
carrega_dados_sinisa(componente, ano)
}
\arguments{
\item{componente}{Componente a ser carregado: "agua", "esgoto" ou "residuos"}
}
\value{
Data frame com os dados do SINISA
}
\description{
Carrega os dados pré-processados do SINISA
}
