% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pnad.R
\name{carrega_pnad_continua}
\alias{carrega_pnad_continua}
\title{Carrega PNAD continua}
\usage{
carrega_pnad_continua(componente, ano)
}
\arguments{
\item{componente}{tema a ser adicionado}

\item{ano}{ano do pnad}
}
\value{
tabela com os dados de atendimento
}
\description{
Carrega PNAD continua
}
