% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R
\name{carrega_sinapi}
\alias{carrega_sinapi}
\title{Carrega dados do SINAPI}
\usage{
carrega_sinapi(nome)
}
\arguments{
\item{nome}{Nome do conjunto de dados retornado pelo lista_sinapi}
}
\value{
um \code{data.frame} com os dados do sinapi
}
\description{
Carrega dados do SINAPI
}
