% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{censo_2022}
\alias{censo_2022}
\title{Cria conjunto de dados populacionais do censo IBGE 2022}
\usage{
censo_2022()
}
\value{
um \code{data.frame} com a estimativa populacional pelo Censo
}
\description{
Cria conjunto de dados populacionais do censo IBGE 2022
}
\examples{
censo_2022()
}
