% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{censo_2022_atendimento}
\alias{censo_2022_atendimento}
\title{Cria conjunto de indicadores domícilos IBGE 2022}
\usage{
censo_2022_atendimento()
}
\value{
um \code{data.frame} com o índice de atendimento de água, esgoto e resíduos sólidos
}
\description{
Cria conjunto de indicadores domícilos IBGE 2022
}
