% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{check_and_create_data_folder}
\alias{check_and_create_data_folder}
\title{Verifica e cria a pasta de armazenamento}
\usage{
check_and_create_data_folder()
}
\description{
Verifica se a pasta de armazenamento existe e caso a pasta não exista ela será criada.
}
\examples{
check_and_create_data_folder()
}
