% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R, R/data_zhelper.R
\name{check_and_create_datasets}
\alias{check_and_create_datasets}
\title{Verifica a integridade e cria amarzenamento de conjunto de dados}
\usage{
check_and_create_datasets()

check_and_create_datasets()
}
\description{
Verifica a integridade e cria amarzenamento de conjunto de dados

Verifica a integridade e cria amarzenamento de conjunto de dados
}
