% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{check_and_create_state}
\alias{check_and_create_state}
\title{Verfica e cria estado da aplicação}
\usage{
check_and_create_state()
}
\value{
estado do aplicativo
}
\description{
Verifica se o estado da aplicação existe e caso não exista cria uma aplicação com valores padrões.
}
\examples{
app_state <- check_and_create_state()
}
