% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{check_data_exists}
\alias{check_data_exists}
\title{check dataset file exists}
\usage{
check_data_exists(name)
}
\arguments{
\item{name}{The variable name}
}
\value{
If file exists (boolean)
}
\description{
check dataset file exists
}
\examples{
check_data_exists("populacao")
}
