% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{classifica_deficit_setor}
\alias{classifica_deficit_setor}
\title{Classifica défict no setor censitário}
\usage{
classifica_deficit_setor(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna \code{populacao}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{classe_densidade}.
}
\description{
Classifica setor de censitário baseando-se no numero de moradores
}
