% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{classifica_densidade_setor}
\alias{classifica_densidade_setor}
\title{Classifica densidade populacional no setor censitário}
\usage{
classifica_densidade_setor(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna \code{densidade_setor}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{classe_densidade}.
}
\description{
Classifica setor censitário baseando-se na densidade populacional.
}
