% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{classifica_faixa_populacional}
\alias{classifica_faixa_populacional}
\title{Faixa populaciona;}
\usage{
classifica_faixa_populacional(
  tabela,
  limites,
  campo_populacao = "POP_TOT",
  campo_faixa = "faixa"
)
}
\arguments{
\item{tabela}{tabela contendo a colunas de população total.}

\item{limites}{limites para faixa populacional. Ex: limites<-c(0,1,2) cria as seguintes faixas:
\itemize{
\item{faixa 0}{pop<0}
\item{faixa 1}{0<=pop<=1}
\item{faixa 2}{0<pop<=2}
\item{faixa 3}{pop>2}
}}

\item{campo_populacao}{parâmetro opcional para definir o campo que contém a populacao (default: "POP_TOT")}

\item{campo_faixa}{parâmetro opcional para definir o campo que contém a populacao (default: "faixa")}
}
\value{
tabela contendo um coluna adcional com as faixas populacionais
}
\description{
Cria classificação por limites populacionas.
}
\examples{
limites <- as.integer(c(0, 10e3, 30e3, 100e3, 250e3, 1e6, 4e6))
POP_TOT <- as.integer(c(0, 10e3, 30e3, 100e3, 250e3, 1e6, 4e6, 5e6))
input <- dplyr::tibble(POP_TOT)
output <- classifica_faixa_populacional(input, limites)
}
