% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{classifica_municipio}
\alias{classifica_municipio}
\title{Classifica município em relação ao seu tamanho e custo de aquisição de materiais}
\usage{
classifica_municipio(tabela)
}
\arguments{
\item{tabela}{Um dataframe contendo as colunas: populacao_total, populacao_rural e populacao_urbana}
}
\value{
Dataframe contendo a coluna adicional \code{classificacao} e \code{cenario}
}
\description{
A classificação é composta de 3 digitos (XXY)
os dois primeiro (XX) é a classificação do tamanho da população e o último (y) refere-se ao custo.
A classificação do tamanho da população é:
\itemize{
\item{07}{ - populações menores ou igual a 40 000 habitantes}
\item{08}{ - populações maiores que 40 000 habitantes e menores ou igual a 400 000 habitantes}
\item{09}{ - populações maiores que 400 000 habitantes}
}
Quanto ao custo:
\itemize{
\item{a}{ - indica alto custo}
\item{b}{ - indica baixo custo}
}
Exemplo: população 07b é um município com população menor que 40000 e com baixo custo para aquisição de materiais e serviços
}
\examples{
populacao_total =  c(3000, 3000, 300000, 300000, 600000, 600000)
populacao_urbana = c(2999,    1, 299999,      1, 599999,      1)
populacao_rural =  c(   1, 2999,      1, 299999,      1, 599999)
tabela <- dplyr::tibble(populacao_total, populacao_urbana, populacao_rural)
tabela <- classifica_municipio(tabela)
}
