% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_zhelper.R
\name{clean_dataset}
\alias{clean_dataset}
\title{Limpa o armazenamendo do conjunto de dados}
\usage{
clean_dataset(name)
}
\arguments{
\item{name}{é um \code{character} com o nome do conjunto de dados}
}
\description{
Limpa o armazenamendo do conjunto de dados
}
