% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{clean_populacao}
\alias{clean_populacao}
\title{Limpa armazenamento de dados populacionais}
\usage{
clean_populacao()
}
\value{
nothing
}
\description{
Limpa armazenamento de dados populacionais
}
\examples{
\dontrun{
clean_populacao()
}
}
