% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localidades.R
\name{codigo6_para_codigo_ibge}
\alias{codigo6_para_codigo_ibge}
\title{Converte codigo de munícipio com 6digitos para codigo completo do IBGE}
\usage{
codigo6_para_codigo_ibge(tabela, nome_campo)
}
\arguments{
\item{tabela}{contendo codigo de 6 digitos}

\item{nome_campo}{nome do campo que contém o codigo de 6 digitos}
}
\value{
uma tabela contendo o campo codigo_municipio (código IBGE)
}
\description{
Converte codigo de munícipio com 6digitos para codigo completo do IBGE
}
\examples{
codes <- c("110001", "110002")
tabela <- dplyr::tibble(codes)
tabela <- codigo6_para_codigo_ibge(tabela, "codes")
}
