% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{codigo_setor_para_municipio}
\alias{codigo_setor_para_municipio}
\title{Código setor para codigo do município}
\usage{
codigo_setor_para_municipio(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna \code{codigo_setor}}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{codigo_municipio}
}
\description{
Transforma os códigos do setor censitário do IBGE para o código de município (IBGE).
}
