% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{coeficiente_pd}
\alias{coeficiente_pd}
\title{Calcula o coeficiente PD}
\usage{
coeficiente_pd(tabela)
}
\arguments{
\item{tabela}{é um \code{data.frame} contendo as colunas \code{precipitacao_moda} e \verb{densidade urbana}}
}
\value{
um \code{data.frame} contendo a coluna \code{pd}
}
\description{
O coeficiente PD é a multiplicação entre a moda da precipitação(mm) e a densidadade urbana (hab/km²) dividos por 1000.
}
\examples{
tabela <- dplyr::tibble(precipitacao_moda = c(1, 1), densidade_urbana = c(2, 2))
tabela <- coeficiente_pd(tabela)

}
