% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{compara_state_json}
\alias{compara_state_json}
\title{Compara o estado atual com o estado salvo
Retorna \code{TRUE} se o estado atual é igual ao estado salvo no arquivo JSON}
\usage{
compara_state_json(state)
}
\arguments{
\item{state}{estrutura de dados (\code{list}) que guarda o estado atual da aplicação}
}
\value{
\code{TRUE} se o estado atual é igual ao estado salvo, \code{FALSE} caso contrário
}
\description{
Compara o estado atual com o estado salvo
Retorna \code{TRUE} se o estado atual é igual ao estado salvo no arquivo JSON
}
