% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{consolida_investimentos}
\alias{consolida_investimentos}
\title{Consolida as necessidade de investimento em saneamento}
\usage{
consolida_investimentos(state)
}
\arguments{
\item{state}{estrutura de dados (\code{list}) que guarda o estado atual da aplicação}
}
\value{
estado da aplicação (\code{list}) que guarda todos os resultado e parâmetros dos cálculos
}
\description{
Consolida as necessidade de investimento em saneamento
}
\examples{
\dontrun{
state <- consolida_investimentos(state)
}
}
