% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{consolida_investimentos_agua}
\alias{consolida_investimentos_agua}
\title{Consolida os dados de investimentos para água}
\usage{
consolida_investimentos_agua(tabela)
}
\arguments{
\item{tabela}{contendo as colunas:
\itemize{
\item{custo_expansao_distribuicao_agua}
\item{custo_expansao_producao_agua}
\item{custo_reposicao_producao_agua}
\item{custo_reposicao_distribuicao_agua}
}}
}
\value{
tabela contendo os campos:
\itemize{
\item{investimento_expansao}
\item{investimento_reposicao}
\item{investimento_total}
}
}
\description{
Totaliza os dados investimento em expansão, reposição e total.
Adiciona coluna com o país para totalização.
}
\examples{
\dontrun{
tabela <- consolida_investimentos_agua(tabela)
}
}
