% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidacao.R
\name{consolida_investimentos_tabela_longa}
\alias{consolida_investimentos_tabela_longa}
\title{Consolida as necessidade de investimento em saneamento em uma tabela longa}
\usage{
consolida_investimentos_tabela_longa(state)
}
\arguments{
\item{state}{estrutura de dados (\code{list}) que guarda o estado atual da aplicação}
}
\value{
estado da aplicação (\code{list}) que guarda todos os resultado e parâmetros dos cálculos
}
\description{
Consolida as necessidade de investimento em saneamento em uma tabela longa
}
\examples{
\dontrun{
state <- consolida_investimentos(state)
}
}
