% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{consolida_populacao_snis}
\alias{consolida_populacao_snis}
\title{Junta as estimativas de populacao total, urbana, rural e dados de snis}
\usage{
consolida_populacao_snis(populacao, ano, snis)
}
\arguments{
\item{populacao}{é os dados da projeção populacional}

\item{ano}{da estimativa de população}

\item{snis}{dados do snis}
}
\value{
tabela com todos os dados consolidados
}
\description{
Junta as estimativas de populacao total, urbana, rural e dados de snis
}
\examples{
\dontrun{
df <- consolida_populacao_snis(fonte1, fonte2, snis2020)
}
}
