% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{corrige_plano_drenagem}
\alias{corrige_plano_drenagem}
\title{Corrige preços do plano de drenagem}
\usage{
corrige_plano_drenagem(data)
}
\arguments{
\item{data}{um \code{Date} contendo a data para correção do preço}
}
\description{
Realiza a correção dos preços para data passada.
}
\examples{
\dontrun{

}
}
