% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_zhelper.R
\name{create_dataset}
\alias{create_dataset}
\title{Cria um dataset}
\usage{
create_dataset(name)
}
\arguments{
\item{name}{é um \code{character} com o nome do conjunto de dados}
}
\description{
Cria um dataset
}
