% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_municipio.R
\name{create_municipio}
\alias{create_municipio}
\title{Cria dataset com informações gerais
sobre os munícipios do Brasil
regiões, estado e códigos do IBGE}
\usage{
create_municipio()
}
\description{
Cria dataset com informações gerais
sobre os munícipios do Brasil
regiões, estado e códigos do IBGE
}
\examples{
create_municipio()
}
