% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{create_populacao}
\alias{create_populacao}
\title{Cria armazenamento local de dados populacionais}
\usage{
create_populacao()
}
\value{
nada
}
\description{
Cria armazenamento local de dados populacionais
}
\examples{
\dontrun{
create_populacao()
}
}
