% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge_load.R
\name{create_populacoes}
\alias{create_populacoes}
\title{Baixa e salva os dados de população}
\usage{
create_populacoes()
}
\value{
1 quando ok, NULL quando falha
}
\description{
Baixa e salva os dados de população
}
\examples{
\dontrun{
create_populacoes()
}
}
