% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{cria_armazenamento_padrao}
\alias{cria_armazenamento_padrao}
\title{Transforma datasets da biblioteca em xlsx}
\usage{
cria_armazenamento_padrao()
}
\description{
Transforma datasets da biblioteca em xlsx
}
