% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cria_faixas_vazias}
\alias{cria_faixas_vazias}
\title{Cria faixa populacional nula em grupos}
\usage{
cria_faixas_vazias(tabela, nome_grupo = "estado")
}
\arguments{
\item{tabela}{tabela contendo as colunas de população, estado}

\item{nome_grupo}{o agrupando dentro de uma faixa populacional (padrão: "estado")}
}
\value{
tabela contendo a média de todas variáveis por faixa populacional
}
\description{
É criada uma faixa com valores vazio (NA) quando não existe uma faixa populacional em um grupo.
}
