% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{custo_coleta_esgoto}
\alias{custo_coleta_esgoto}
\title{Custo coletivos para sistemas de coleta de esgoto}
\usage{
custo_coleta_esgoto(tabela, coleta_esgoto)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{seguranca_hidrica} e \code{semiarido}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{custo_agua_individual}.
}
\description{
Adiciona coluna com custo coletivos para sistemas de coleta de esgoto.
}
