% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{custo_distribuicao_agua}
\alias{custo_distribuicao_agua}
\title{Custo coletivos para sistemas de abastecimento de água}
\usage{
custo_distribuicao_agua(tabela, custo_distribuicao)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{seguranca_hidrica} e \code{semiarido}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{custo_agua_individual}.
}
\description{
Adiciona coluna com custo coletivos para sistemas de abastecimento de água.
}
