% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{custo_individual_esgoto}
\alias{custo_individual_esgoto}
\title{Custo individual para sistemas de estações de esgoto}
\usage{
custo_individual_esgoto(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{seguranca_hidrica} e \code{V003}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{custo_esgoto_individual }.
}
\description{
Adiciona coluna com custo individual para sistemas de estações de esgoto.
}
