% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_zhelper.R
\name{dataset_exists}
\alias{dataset_exists}
\title{Verifica se o conjunto de dados existe}
\usage{
dataset_exists(name)
}
\arguments{
\item{name}{é um \code{character} com o nome do conjunto de dados}
}
\value{
\code{logical} sendo TRUE conjunto de dados encontrado e FALSE conjunto inexistente
}
\description{
Verifica se o conjunto de dados existe
}
