% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{deficit_coleta_seletiva}
\alias{deficit_coleta_seletiva}
\title{Déficit de coleta seletiva}
\usage{
deficit_coleta_seletiva(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{populacao_urbana} e \code{atendimento_relativo_seletiva_urbano}}
}
\value{
um \code{data.frame} contendo a coluna deficit_coleta_seletiva
}
\description{
Déficit de coleta seletiva
}
