% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{demanda_aterro}
\alias{demanda_aterro}
\title{Demanda por aterro}
\usage{
demanda_aterro(tabela, vida_util)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna \code{total_residuos_projecao}}

\item{vida_util}{valor em anos da vida util do aterro}
}
\value{
um \code{data.frame} contendo a coluna \code{demanda_aterro}
}
\description{
Demanda por aterro
}
