% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{demanda_compostagem}
\alias{demanda_compostagem}
\title{Demanda por compostagem}
\usage{
demanda_compostagem(tabela)
}
\arguments{
\item{tabela}{}
}
\description{
Demanda por compostagem
}
\examples{
\dontrun{

}
}
