% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{demanda_transbordo}
\alias{demanda_transbordo}
\title{Demanda por transbordo}
\usage{
demanda_transbordo(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo numero_municipio}
}
\value{
um \code{data.frame} contendo a coluna \code{demanda_tranbordo}
}
\description{
A demanda por transbordo é considerada 1 estacao de transbordo e 1 caminhao por municipio
}
\examples{
\dontrun{

}
}
