% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{demanda_triagem}
\alias{demanda_triagem}
\title{Demanda por triagem}
\usage{
demanda_triagem(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{meta_reaproveitamento}, \code{total_residuos_projecao} e \code{CS026}}
}
\value{
um \code{data.frame} com a coluna \code{demanda_triagem}
}
\description{
Demanda por triagem
}
