% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{densidade_caminhao_por_estado_faixa}
\alias{densidade_caminhao_por_estado_faixa}
\title{Preenche caminhao}
\usage{
densidade_caminhao_por_estado_faixa(
  tabela_faixa,
  tabela_por_municipio,
  metodo = "media"
)
}
\arguments{
\item{tabela_por_municipio}{um \code{data.frame} com as colunas "codigo_municipio", "faixa", "CS001", "CS050", "CO064", "CO065", "CO066", "CO067", "CO068"}

\item{estado_faixa}{um \code{data.frame} com as colunas "estado", "faixa", "POP_URB", "CS050", "CS001"}
}
\value{
um \code{data.frame} contendo a coluna atendimento_relativo_coleta_seletiva
}
\description{
\enumerate{
\item Calcula a quantidade de caminhões por faixa
}
}
