% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{densidade_caminhoes}
\alias{densidade_caminhoes}
\title{Densidade de caminhões}
\usage{
densidade_caminhoes(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas CO164 e numero_caminhoes}
}
\value{
um \code{data.frame} contendo a coluna \code{densidade_caminhoes}
}
\description{
Cálcula a densidade de caminhões
}
