% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{densidade_caminhoes_compactadores}
\alias{densidade_caminhoes_compactadores}
\title{Densidade de caminhões compactadores}
\usage{
densidade_caminhoes_compactadores(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas atendimento_coleta_indiferenciada_hab e quantidade_caminhoes_compactadores}
}
\value{
um \code{data.frame} contendo a coluna \code{densidade_caminhoes_compactadores}
}
\description{
Cálcula a densidade de caminhões compactadores
}
