% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{densidade_setor}
\alias{densidade_setor}
\title{Densidade populacional no setor censitário}
\usage{
densidade_setor(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{V002_projecao} e \code{area_setor_km2}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{densidade_setor}.
}
\description{
Cálcula densidade populacional no setor (V002/area_setor_km2)
}
