% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{densidade_urbana}
\alias{densidade_urbana}
\title{Densidade urbana}
\usage{
densidade_urbana(tabela)
}
\arguments{
\item{tabela}{contendo as colunas do SNIS: \code{GE006} e \code{GE002}}
}
\value{
uma tabela contendo a densidade urbana
}
\description{
Densidade urbana
}
\examples{
\dontrun{
tabela <- densidade_urbana(tabela)
}
}
