% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{disposicao_inadequada}
\alias{disposicao_inadequada}
\title{Adiciona quantidade de residuos com disposição inadequada}
\usage{
disposicao_inadequada(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} com a coluna \code{CO119}}
}
\value{
um \code{data.frame} com a colunas adicional \code{residuos_disposicao_inadequada}
}
\description{
Adiciona quantidade de residuos com disposição inadequada
}
