% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{divide_residuo_litoraneo}
\alias{divide_residuo_litoraneo}
\title{Divide quantidade de residuo litorâneo e nao litoraneo}
\usage{
divide_residuo_litoraneo(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas CO119, CO164 e litoral}
}
\value{
um \code{data.frame} onde os municipios litoraneos não tem dados de quantidade de residuos
}
\description{
Divide quantidade de residuo litorâneo e nao litoraneo
}
\examples{
\dontrun{
tbl <- divide_residuo_litoraneo(tbl)
}
}
