% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{domicilios_adequados_com_agua}
\alias{domicilios_adequados_com_agua}
\title{Número de domicílios sistemas de água adequados}
\usage{
domicilios_adequados_com_agua(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{domicilios_projecao} e \code{deficit_agua_relativo_rural}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{domicilios_adequados_agua}.
}
\description{
Calcula o número de domicílios com sistemas de abastecimento de água adequados.
}
