% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{download_and_preprocess_sinisa}
\alias{download_and_preprocess_sinisa}
\title{Download SINISA and preprocess data}
\usage{
download_and_preprocess_sinisa(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\description{
Download SINISA and preprocess data
}
