% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{download_estimativa_populacao}
\alias{download_estimativa_populacao}
\title{Cria conjunto de dados de estimativas populacionais por município do IBGE}
\usage{
download_estimativa_populacao()
}
\value{
nada
}
\description{
Cria conjunto de dados de estimativas populacionais por município do IBGE
}
\examples{
download_estimativa_populacao()
}
