% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R
\name{download_sinapi}
\alias{download_sinapi}
\title{Baixa dados do SINAPI para um determinado ano e mês}
\usage{
download_sinapi(year, month)
}
\arguments{
\item{year}{é um \code{character} com o ano desejado}

\item{month}{é um \code{character} com o mês desejado}
}
\value{
um \code{data.frame} com os dados do SINAPI
}
\description{
Baixa dados do SINAPI para um determinado ano e mês
}
