% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R
\name{download_sinapi_v2025}
\alias{download_sinapi_v2025}
\title{Download SINAPI 2025
Salva dados brutos do sinapi}
\usage{
download_sinapi_v2025(ano, mes)
}
\arguments{
\item{ano}{um \code{number} com o ano desejado}

\item{mes}{um \code{number} com o mês desejado}
}
\value{
filename do arquivo baixado
}
\description{
Download SINAPI 2025
Salva dados brutos do sinapi
}
