% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{download_sinisa}
\alias{download_sinisa}
\title{Download SINISA data}
\usage{
download_sinisa(year)
}
\arguments{
\item{year}{Year of the data to download (e.g., 2022)}
}
\value{
Path to the downloaded files
}
\description{
Download SINISA data
}
