% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{download_sinisa_raw}
\alias{download_sinisa_raw}
\title{Download SINISA raw data}
\usage{
download_sinisa_raw(year)
}
\arguments{
\item{year}{Year of the data to download (e.g., 2022)}
}
\value{
Path to the downloaded files
}
\description{
Download SINISA raw data
}
