% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis_ap.R
\name{download_snis_ap}
\alias{download_snis_ap}
\title{Retorna dados do SNIS-AP}
\usage{
download_snis_ap(year)
}
\arguments{
\item{year}{ano desejado}
}
\value{
um \code{data.frame} contendo os dados do SNIS
}
\description{
Retorna dados do SNIS-AP
}
