% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis_rs.R
\name{download_snis_rs}
\alias{download_snis_rs}
\title{Retorna dados do SNIS-RS}
\usage{
download_snis_rs(year)
}
\arguments{
\item{year}{ano desejado}
}
\value{
um \code{data.frame} contendo os dados do SNIS
}
\description{
Retorna dados do SNIS-RS
}
