% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge_load.R
\name{downloadCensoRawData}
\alias{downloadCensoRawData}
\title{Cria conjunto de dados do censo IBGE}
\usage{
downloadCensoRawData()
}
\value{
Dataframe com informações dos conjuntos de dados baixados
}
\description{
Cria conjunto de dados do censo IBGE
}
\examples{
\dontrun{
downloadCensoRawData()
}
}
