% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge_load.R
\name{downloadEstimativaPopRawData}
\alias{downloadEstimativaPopRawData}
\title{Cria conjunto de dados de estimativas populacionais por município do IBGE}
\usage{
downloadEstimativaPopRawData()
}
\value{
Um data.frame() contendo a lista de estimativas criadas
}
\description{
Cria conjunto de dados de estimativas populacionais por município do IBGE
}
\examples{
\dontrun{
downloadEStimativaPopRawData()
}
}
