% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{esgoto_preprocess}
\alias{esgoto_preprocess}
\title{Esgoto Pre-processing
Read the SINISA sewage data and process it into a data frame.}
\usage{
esgoto_preprocess(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\value{
A data frame containing the processed sewage data
}
\description{
Esgoto Pre-processing
Read the SINISA sewage data and process it into a data frame.
}
