% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto.R
\docType{data}
\name{esgoto_required_fields}
\alias{esgoto_required_fields}
\title{Necessidade de Investimento - Componente Esgoto - Situação Urbana}
\format{
An object of class \code{character} of length 10.
}
\usage{
esgoto_required_fields
}
\description{
Necessidade de Investimento - Componente Esgoto - Situação Urbana
}
\keyword{datasets}
