% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igp.R
\name{fator_correcao_ipea}
\alias{fator_correcao_ipea}
\title{Fator de Correção de Preços - IPEA}
\usage{
fator_correcao_ipea(ipea_code, data_inicial, data_final)
}
\arguments{
\item{ipea_code}{Código do índice no IPEA Data.}

\item{data_inicial}{Um vetor de datas (\code{Date}) contendo as datas iniciais para o cálculo.}

\item{data_final}{Um vetor de datas (\code{Date}) contendo as datas finais para o cálculo.}
}
\value{
Um vetor numérico (\code{double}) com os fatores de correção correspondentes para cada par de datas.
}
\description{
Calcula o fator de correção de preços com base no índice fornecido pelo IPEA para um intervalo de datas.
}
\examples{
data_inicial <- as.Date(c("2016-01-01", "2017-01-01"))
data_final <- as.Date(c("2021-01-01", "2022-01-01"))
fator <- fator_correcao_ipea("IGP12_INCC12", data_inicial, data_final)
}
