% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{fator_correcao_perdas}
\alias{fator_correcao_perdas}
\title{Retorna o fator de correção para as perdas no sistema de abastecimento}
\usage{
fator_correcao_perdas(perda)
}
\arguments{
\item{perda}{é a porcentagem de perda em decimais (25 = 25\%).}
}
\value{
o fator de correção devido a perda no sistema de abastecimento
}
\description{
Retorna o fator de correção para as perdas no sistema de abastecimento
}
\examples{
fator <- fator_correcao_perdas(25)
}
