% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{fator_perda_agua}
\alias{fator_perda_agua}
\title{Fator de perda de água}
\usage{
fator_perda_agua(perda)
}
\arguments{
\item{perda}{é a porcentagem de perda em \%.}
}
\value{
O fator de perda,
}
\description{
O fator de perda de água é valor pelo qual a demanda de água deve ser multiplicada para compensar as perdas.
}
\examples{
fator <- fator_perda_agua(25)
}
