% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{fill_missing_density}
\alias{fill_missing_density}
\title{Completa os dados faltantes de densidade.}
\usage{
fill_missing_density(density)
}
\arguments{
\item{density}{um tibble() contendo os dados de densidades}
}
\value{
um tibble() com a mesma tabela mas sem dados faltantes
}
\description{
Completa os dados faltantes de densidade.
}
\examples{
\dontrun{
df <- fill_missing_density(density)
}
}
