% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{fracao_coletivo_individual_agua}
\alias{fracao_coletivo_individual_agua}
\title{Fração de investimentos em água coletivos e individuais}
\usage{
fracao_coletivo_individual_agua(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{codigo_situacao}, \code{deficit_agua_relativo_rural} e \code{regiao}.}
}
\value{
um \code{data.frame} contendo as colunas adicionais \code{fracao_investimento_coletivo_agua} e \code{fracao_investimento_individual_agua}.
}
\description{
Define a fração coletiva e individual para necessidade de investimento em água.
}
