% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{fracao_coletivo_individual_esgoto}
\alias{fracao_coletivo_individual_esgoto}
\title{Fração de investimentos em esgoto coletivos e individuais}
\usage{
fracao_coletivo_individual_esgoto(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{situacao_setor}, \code{deficit_esgoto_relativo_rural} e \code{regiao}.}
}
\value{
um \code{data.frame} contendo as colunas adicionais \verb{fracao_investimento_coletivo_esgoto e }fracao_investimento_individual_esgoto`.
}
\description{
Define a fração coletiva e individual para necessidade de investimento em esgoto.
}
