% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{geracao_residuos}
\alias{geracao_residuos}
\title{Geração de resíduos}
\usage{
geracao_residuos(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas populacao_total_corrente, residuo_coletado_ton_ano, atendimento_coleta_indiferenciada_hab, residuo_recuperado_ton_ano e populacao_total}
}
\value{
um \code{data.frame} contendo as colunas taxa_geracao_residuos, total_residuos, total_residuos_projecaoe
}
\description{
Geração de resíduos
}
