% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R, R/data_universan.R
\name{get_censo_labels}
\alias{get_censo_labels}
\title{Retorna os rótulos dos dados do IBGE disponíveis (somente censo)}
\usage{
get_censo_labels()

get_censo_labels()
}
\value{
um vetor com os dados disponiveis

um vetor com os dados disponiveis

Um vetor com os nomes dos arquivos de censo disponíveis (sem extensão).
}
\description{
Retorna os rótulos dos dados do IBGE disponíveis (somente censo)

Retorna os rótulos dos dados do IBGE disponíveis (somente censo)
}
\examples{
\dontrun{
df <- get_sinapi_data(sinapi)
}
\dontrun{
df <- get_sinapi_data(sinapi)
}
Retorna os rótulos dos arquivos de censo disponíveis na base de cálculo

Esta função lista os arquivos na pasta 'dados/base_calculo' que contêm o padrão "censo" no nome.

}
