% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{get_censo_years}
\alias{get_censo_years}
\title{Lista de anos de população estimada pelo Censo
que estão disponíveis para download no ftp do IBGE}
\usage{
get_censo_years()
}
\arguments{
\item{url}{}
}
\value{
lista de anos do censo disponíveis
}
\description{
Lista de anos de população estimada pelo Censo
que estão disponíveis para download no ftp do IBGE
}
\examples{
get_censo_years()
}
