% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_data_dir}
\alias{get_data_dir}
\title{get dataset dir}
\usage{
get_data_dir()
}
\value{
The directory where dataset is stored
}
\description{
get dataset dir
}
\examples{
path <- get_data_dir("populacao")
}
