% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_data_path}
\alias{get_data_path}
\title{Retorna o caminho de um conjunto de dados local}
\usage{
get_data_path(name)
}
\arguments{
\item{name}{nome do conjunto de dados}
}
\value{
path to dataset .rda file
}
\description{
Retorna o caminho de um conjunto de dados local
}
\examples{
path <- get_data_path("populacao")
}
