% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametros.R
\name{get_default_input}
\alias{get_default_input}
\title{Valores padrões da aplicação}
\usage{
get_default_input()
}
\value{
estrutura de dados que guarda os valores padrões da interface gráfica
}
\description{
Retorna os valores padrões dos parâmetros de todos os módulos do cálculo de investimento.
}
\examples{
get_default_input()
}
