% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{get_populacao}
\alias{get_populacao}
\title{Ferramentas para Cálculo das Necessidades de Investimento
para o Saneamento Água e Esgoto
Retorna a projecao de população de um tipo e ano}
\usage{
get_populacao(tabela, ano, tipo)
}
\arguments{
\item{tabela}{projecao da populacao}

\item{ano}{é o Ano que se deseja a projeção}

\item{tipo}{é o tipo de população (total, rural ou urbana)}
}
\value{
dados da projeção para um determinado tipo de população
}
\description{
Ferramentas para Cálculo das Necessidades de Investimento
para o Saneamento Água e Esgoto
Retorna a projecao de população de um tipo e ano
}
\examples{
\dontrun{
df <- get_populacao(df, 2033, "urbana")
}
}
