% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R, R/data_universan.R
\name{get_populacao_labels}
\alias{get_populacao_labels}
\title{Retorna os rótulos dos dados do IBGE disponíveis}
\usage{
get_populacao_labels()

get_populacao_labels()
}
\value{
um vetor com os dados disponiveis

um vetor com os dados disponiveis
}
\description{
Retorna os rótulos dos dados do IBGE disponíveis

Retorna os rótulos dos dados do IBGE disponíveis
}
\examples{
\dontrun{
df <- get_sinapi_data(sinapi)
}
\dontrun{
df <- get_sinapi_data(sinapi)
}
}
