% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{get_sinapi_data}
\alias{get_sinapi_data}
\title{Retorna dados do SINAPI}
\usage{
get_sinapi_data(sinapi)
}
\arguments{
\item{sinapi}{caminho do dados do SINAPI}
}
\value{
um tibble() com dados do SINAPI
}
\description{
Retorna dados do SINAPI
}
\examples{
\dontrun{
df <- get_sinapi_data(sinapi)
}
}
