% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis.R
\name{get_snis_data}
\alias{get_snis_data}
\title{Retorna dados do SNIS}
\usage{
get_snis_data(snis, fields)
}
\arguments{
\item{snis}{caminho do SNIS}

\item{fields}{lista com campos a serem retornados}
}
\value{
um \code{data.frame} contendo os dados do SNIS
}
\description{
Retorna dados do SNIS
}
