% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis_rs.R
\name{get_snis_rs_list}
\alias{get_snis_rs_list}
\title{Retorna lista dos anos de SNIS-AP}
\usage{
get_snis_rs_list()
}
\value{
um \code{list} contendo os anos para tentar baixar
}
\description{
Retorna lista dos anos de SNIS-AP
}
