% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igp.R
\name{get_taxa_igp}
\alias{get_taxa_igp}
\title{Calcula a taxa de correção de preço usando o IGP}
\usage{
get_taxa_igp(igp, data_inicial, data_final)
}
\arguments{
\item{igp}{é a tabela do IGP.}

\item{data_inicial}{é um date contendo a data inicial}

\item{data_final}{é um date contendo a data final}
}
\value{
a taxa que deve ser multiplicada pelo valor em R$ para corrígi-lo.
}
\description{
A taxa inclui o mês da data inicial
}
\examples{
data(igp)
valor <- function(igp, as.Date("2016-01-01"), as.Date("2026-01-01"))
}
