% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{get_year_from_path}
\alias{get_year_from_path}
\title{Retorna o ano do dataset de população}
\usage{
get_year_from_path(path)
}
\arguments{
\item{path}{Caminho do dataset de população}
}
\value{
ano do dataset de populacao
}
\description{
Retorna o ano do dataset de população
}
\examples{
ano <- get_year_from_path("populaca_estimada_2021")
}
