% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge_load.R
\name{getCensoYears}
\alias{getCensoYears}
\title{Lista de anos de população estimada pelo Censo
que estão disponíveis para download no ftp do IBGE}
\usage{
getCensoYears()
}
\arguments{
\item{URL}{}
}
\value{
lista de anos do censo disponíveis
}
\description{
Lista de anos de população estimada pelo Censo
que estão disponíveis para download no ftp do IBGE
}
\examples{
\dontrun{
getCensoYears()
}
}
