% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{habitantes_com_deficit_esgoto}
\alias{habitantes_com_deficit_esgoto}
\title{Número de habitantes com deficit de esgoto}
\usage{
habitantes_com_deficit_esgoto(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{domicilios_projecao} e \code{deficit_esgoto_relativo_rural}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{domicilios_deficit_esgoto}.
}
\description{
Calcula o número de habitantes com déficit de sistemas de estação de esgoto.
}
