% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{ibge_id_to_name}
\alias{ibge_id_to_name}
\title{Transforma id IBGE em nome legível}
\usage{
ibge_id_to_name(id)
}
\value{
um \code{character} com o nome
}
\description{
Transforma id IBGE em nome legível
}
