% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{indicador_cobertura_rede_agua}
\alias{indicador_cobertura_rede_agua}
\title{Cria indicador de cobertura de rede de água
indicador = tem_rede /populacao}
\usage{
indicador_cobertura_rede_agua(tabela)
}
\arguments{
\item{tabela}{é um \code{data.frame} com os dados agregados por setor censitário da população}
}
\value{
um \code{data.frame} com os dados agregados por município e o indicador de cobertura de rede de água
}
\description{
Cria indicador de cobertura de rede de água
indicador = tem_rede /populacao
}
