% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indices_drenagem}
\alias{indices_drenagem}
\title{Índices de Drenagem Urbana}
\format{
Uma dataframe com um muncípio por linha e as colunas:
\describe{
\item{codigo_municipio}{Código IBGE do município}
\item{caracteristicas_fisicas}{índice de caracteristica físicas}
\item{infraestrutura}{índice de infraestrutura}
}
}
\usage{
indices_drenagem
}
\description{
Dataset contendo os índices de caracteristícas físicas e infraestrutura
}
\keyword{datasets}
