% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_municipio.R
\name{integrity_municipio}
\alias{integrity_municipio}
\title{Verifica a integridade dos dados locais de munícipio}
\usage{
integrity_municipio()
}
\description{
Verifica a integridade dos dados locais de munícipio
}
\examples{
is_ok <- integrity_muncipio()
}
