% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{integrity_populacao}
\alias{integrity_populacao}
\title{Verifica a integridade do armazenamento local de dados de população}
\usage{
integrity_populacao()
}
\value{
boleano indicando integridade ok ou não
}
\description{
Verifica a integridade do armazenamento local de dados de população
}
\examples{
integrity_populacao()
}
