% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_projeto.R
\name{integrity_projeto}
\alias{integrity_projeto}
\title{Verifica integridade dos dados dos projetos tipos}
\usage{
integrity_projeto()
}
\value{
um \code{logical} sendo \code{TRUE} integridade OK.
}
\description{
Verifica integridade dos dados dos projetos tipos
}
