% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R
\name{integrity_sinapi}
\alias{integrity_sinapi}
\title{Armazena dados sinapi}
\usage{
integrity_sinapi()
}
\value{
um \code{logical} sendo \code{TRUE} integridade OK.
}
\description{
Armazena dados sinapi
}
