% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis.R
\name{integrity_snis}
\alias{integrity_snis}
\title{Verifica a integridade dos dados do SNIS}
\usage{
integrity_snis()
}
\value{
um \code{logical} sendo \code{TRUE} integridade OK.
}
\description{
Verifica a integridade dos dados do SNIS
}
