% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis_ap.R
\name{integrity_snis_ap}
\alias{integrity_snis_ap}
\title{Verifica a integridade dos dados do SNIS-AP}
\usage{
integrity_snis_ap()
}
\value{
um \code{logical} sendo \code{TRUE} integridade OK.
}
\description{
Verifica a integridade dos dados do SNIS-AP
}
