% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{investimento_agua}
\alias{investimento_agua}
\title{Calcula a necessidade de investimento para água}
\usage{
investimento_agua(state)
}
\arguments{
\item{state}{estado da aplicação}
}
\value{
estado da aplicação
}
\description{
Calcula a necessidade de investimento para água
}
\examples{
\dontrun{
app_state <- investimento_agua(state)
}
}
