% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{investimento_agua_esgoto}
\alias{investimento_agua_esgoto}
\title{Title}
\usage{
investimento_agua_esgoto(state)
}
\arguments{
\item{state}{}
}
\description{
Title
}
\examples{
\dontrun{
investimento_agua_esgoto(state)
}
}
