% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{investimento_cadastro}
\alias{investimento_cadastro}
\title{Investimento em Cadastro Técnico}
\usage{
investimento_cadastro(tabela, valor)
}
\arguments{
\item{tabela}{contendo a coluna código do município}
}
\value{
a tabela de entrada com os campos adicionais dos índices (\code{caracteriticas_fisicas} e \code{infraestrutura})
}
\description{
Calcula o valor necessário para investir em cadastro técnico.
}
