% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{investimento_constante}
\alias{investimento_constante}
\title{Investimento constante}
\usage{
investimento_constante(tabela, valor)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a projeção da população urbana}

\item{valor}{um \code{double} contendo o custo de expansão per capita (R$/hab)}
}
\value{
tabela com os valores de investimento
}
\description{
Valor
}
