% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto.R
\name{investimento_esgoto}
\alias{investimento_esgoto}
\title{Calcula a necessidade de investimento para esgoto}
\usage{
investimento_esgoto(state)
}
\arguments{
\item{state}{estado da aplicação}
}
\value{
estado da aplicação
}
\description{
Calcula a necessidade de investimento para esgoto
}
\examples{
\dontrun{
app_state <- investimento_esgoto(state)
}
}
