% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{investimento_expansao_aterro}
\alias{investimento_expansao_aterro}
\title{Investimento em Aterros}
\usage{
investimento_expansao_aterro(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{demanda_aterro} e \code{preco_unidade_aterro}}
}
\value{
um \code{data.frame} contendo a coluna \code{investimento_expansao_aterro}
}
\description{
Estima o investimento necessários em aterro como a multiplicação de: total de residuos, preço unitario por massa e vida útil do aterro.
}
