% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{investimento_expansao_coleta_indiferenciada}
\alias{investimento_expansao_coleta_indiferenciada}
\title{Necessidade de investimento em coleta indiferenciada}
\usage{
investimento_expansao_coleta_indiferenciada(tabela, valor)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{deficit_coleta_indiferenciada} e \code{densidade_caminhoes}}

\item{valor}{um \code{double} contendo o preço de um caminhão}
}
\value{
um \code{data.frame} contendo a coluna \code{investimento_expansao_coleta_indiferenciada}
}
\description{
Necessidade de investimento em coleta indiferenciada
}
