% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{investimento_expansao_coleta_seletiva}
\alias{investimento_expansao_coleta_seletiva}
\title{Necessidade de investimento em coleta seletiva}
\usage{
investimento_expansao_coleta_seletiva(tabela, valor)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{deficit_coleta_seletiva} e \code{densidade_caminhoes_bau}}

\item{valor}{um \code{double} contendo o preço de um caminhão bau}
}
\value{
um \code{data.frame} contendo a coluna \code{investimento_expansao_coleta_seletiva}
}
\description{
Necessidade de investimento em coleta seletiva
}
