% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{investimento_expansao_compostagem}
\alias{investimento_expansao_compostagem}
\title{Investimento em compostagem}
\usage{
investimento_expansao_compostagem(tabela, vida_util)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{demanda_compostagem} e \code{preco_unidade_compostagem}}

\item{vida_util}{um \code{number} com o numero de anos da vida útil do sistema}
}
\value{
um \code{data.frame} com a coluna \code{investimento_expansao_compostagem}
}
\description{
Investimento em compostagem
}
