% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{investimento_expansao_transbordo}
\alias{investimento_expansao_transbordo}
\title{Investimento em transbordo}
\usage{
investimento_expansao_transbordo(tabela, custo)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna \code{demanda_transbordo}}

\item{vida_util}{valor em anos da vida util do aterro}
}
\value{
um \code{data.frame} contendo a coluna \code{investimento_expansao_transbordo}
}
\description{
Estima o investimento necessários em transbordo como a multiplicação da demanda pelo custo.
}
