% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{investimento_residuos_total}
\alias{investimento_residuos_total}
\title{Totaliza os valores de investimentos em resíduos}
\usage{
investimento_residuos_total(tabela)
}
\arguments{
\item{tabela}{contendo os valores de reposição e investimento em resíduos:
\itemize{
\item\link{investimento_reposicao_aterro}
\item\link{investimento_reposicao_compostagem}
\item\link{investimento_reposicao_triagem}
\item\link{investimento_reposicao_coleta_indiferenciada}
\item\link{investimento_reposicao_coleta_seletiva}
\item\link{investimento_expansao_aterro}
\item\link{investimento_expansao_compostagem}
\item\link{investimento_expansao_triagem}
\item\link{investimento_expansao_transbordo}
\item\link{investimento_expansao_coleta_indiferenciada}
\item\link{investimento_expansao_coleta_seletiva}
}}
}
\value{
um \code{data.frame} contendo as colunas adicionais \code{investimento_reposicao}, \code{investimento_expansao} e \code{investimento_total}
}
\description{
Consolida os valores de investimentos para reposicao, expansão e total
}
