% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{investimento_rural_agua}
\alias{investimento_rural_agua}
\title{Necessidade de investimento rural para sistema de abastecimento de água}
\usage{
investimento_rural_agua(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas:
\code{custo_agua_individual}, \code{custo_agua_producao_rural}, \code{custo_agua_distribuicao_rural},
\code{fracao_investimento_individual_agua}, \code{fracao_investimento_coletivo_agua} e
\code{domicilios_deficit_agua}}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{domicilios_adequados_agua}.
}
\description{
Calcula a necessidade de investimento para expansão do sistema de abastecimento de água.
}
