% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{investimento_total}
\alias{investimento_total}
\title{Investimento total}
\usage{
investimento_total(tabela)
}
\arguments{
\item{tabela}{contendo os campos \code{investimento_expansao} e \code{investimento_reposicao}}
}
\value{
tabela com a soma dos investimento no campo \code{investimento_total}
}
\description{
Soma os investimentos de expansao e resposicao
}
\examples{
\dontrun{

}
}
