% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{investimento_total_drenagem}
\alias{investimento_total_drenagem}
\title{Necessidade de Investimento - Componente Drenagem - Situação Urbana}
\usage{
investimento_total_drenagem(tabela)
}
\arguments{
\item{tabela}{contendo os campos \code{investimento_expansao} e \code{investimento_reposicao}}
}
\value{
tabela com a soma dos investimento no campo \code{investimento_total}
}
\description{
Investimento total
}
\details{
Soma os investimentos de expansao e resposicao
}
