% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{investimento_triagem}
\alias{investimento_triagem}
\title{Investimento em triagem}
\usage{
investimento_triagem(tabela, vida_util)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{demanda_triagem} e \code{preco_unidade_triagem}}

\item{vida_util}{um \code{number} com o numero de anos da vida útil do sistema de triagem}
}
\value{
um \code{data.frame} com a coluna \code{investimento_triagem}
}
\description{
Investimento em triagem
}
