% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{junta_fontes_populacao}
\alias{junta_fontes_populacao}
\title{Junta as duas fontes de população.
São adicionados os sufixos _fonte1 e _fonte2 nos campos com mesmo nome.}
\usage{
junta_fontes_populacao(fonte1, fonte2)
}
\arguments{
\item{fonte1}{é a fonte de dados de população mais antiga.}

\item{fonte2}{é a fonte de dados de população mais recente.}
}
\value{
Um tibble() contendo as duas fontes unidas pelo código do munícipio.
}
\description{
Junta as duas fontes de população.
São adicionados os sufixos _fonte1 e _fonte2 nos campos com mesmo nome.
}
\examples{
\dontrun{
df3 <- junta_fontes_populacao(df1, df2)
}
}
