% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge_load.R
\name{listFilesFromFTP}
\alias{listFilesFromFTP}
\title{List of files/folders in a FTP folder}
\usage{
listFilesFromFTP(URL)
}
\arguments{
\item{URL}{}
}
\value{
list of files and folders
}
\description{
List of files/folders in a FTP folder
}
\examples{
\dontrun{
listFilesFromFTP("ftp.ibge.gov.br")
listFilesFromFTP("ftp.ibge.gov.br/Censos/")
}
}
