% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{load_app_state}
\alias{load_app_state}
\title{Carrega o estado do aplicativo
caso nunca tenha sido aberto
cria um estado padrão}
\usage{
load_app_state()
}
\value{
estado do aplicativo
}
\description{
Carrega o estado do aplicativo
caso nunca tenha sido aberto
cria um estado padrão
}
\examples{
app_state <- load_app_state()
}
