% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{load_data}
\alias{load_data}
\title{load app dataset}
\usage{
load_data(name)
}
\arguments{
\item{name}{The variable name}
}
\value{
the content of the dataset
}
\description{
load app dataset
}
\examples{
populacao <- load_data("populacao")
}
