% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge_load.R
\name{load_estimativa_years}
\alias{load_estimativa_years}
\title{Lista de anos disponíveis de estimativas de população por amostra do IBGE}
\usage{
load_estimativa_years()
}
\value{
Lista de anos disponíveis
}
\description{
Lista de anos disponíveis de estimativas de população por amostra do IBGE
}
\examples{
\dontrun{
load_estimativa_years()
}
}
