% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{load_ibge}
\alias{load_ibge}
\title{Carrega conjunto de dados do IBGE}
\usage{
load_ibge(id)
}
\arguments{
\item{id}{um \code{character} com o id}
}
\value{
o conjunto de dados do SINAPI
}
\description{
Carrega conjunto de dados do IBGE
}
