% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R
\name{load_sinapi}
\alias{load_sinapi}
\title{Carrega conjunto de dados do SINAPI}
\usage{
load_sinapi(id)
}
\arguments{
\item{id}{um \code{character} com o id}
}
\value{
o conjunto de dados do SINAPI
}
\description{
Carrega conjunto de dados do SINAPI
}
