% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis_ap.R
\name{load_snis_ap}
\alias{load_snis_ap}
\title{Carrega conjunto de dados do SNIS-AP}
\usage{
load_snis_ap(id)
}
\arguments{
\item{id}{um \code{character} com o id}
}
\value{
o conjunto de dados do SNIS
}
\description{
Carrega conjunto de dados do SNIS-AP
}
