% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis_rs.R
\name{load_snis_rs}
\alias{load_snis_rs}
\title{Carrega conjunto de dados do SNIS-RS}
\usage{
load_snis_rs(id)
}
\arguments{
\item{id}{um \code{character} com o id}
}
\value{
o conjunto de dados do SNIS
}
\description{
Carrega conjunto de dados do SNIS-RS
}
