% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{media_por_estado_faixa}
\alias{media_por_estado_faixa}
\title{Média de todas as variáveis numéricas agrupando por estado e faixa populacional}
\usage{
media_por_estado_faixa(tabela, campo_estado = "Estado", campo_faixa = "faixa")
}
\arguments{
\item{tabela}{tabela contendo as colunas de população, estado}

\item{campo_estado}{parâmetro opcional para definir o campo que
contém a sigla do estado (default: "Estado")}

\item{campo_faixa}{parâmetro opcional para definir o campo que
contém a faixa populaciona (default: "faixa")}
}
\value{
tabela contendo a média de todas variáveis
por grupos de estado e faixa populacional
}
\description{
Média de todas as variáveis numéricas agrupando por estado e faixa populacional
}
\examples{
Estado <- c(rep("AC", 4), rep("AL", 4))
faixa <- c(rep(1, 2), rep(3, 2), rep(2, 2), rep(4, 2))
valor <- c(1, 1, 2, 2, 3, 3, 4, 4)
input <- dplyr::tibble(Estado, faixa, valor)
output <- media_por_estado_faixa(input)
}
