% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{meta_plansab_residuo}
\alias{meta_plansab_residuo}
\title{Adiciona metas de atendimento por regio}
\usage{
meta_plansab_residuo(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} com a coluna \code{regiao}}
}
\value{
um \code{data.frame} com as colunas adicionais \code{meta_coleta}, \code{meta_compostagem}e \code{meta_reaproveitamento}
}
\description{
Adiciona metas de atendimento por regio
}
