% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipio}
\alias{municipio}
\title{Municípios do Brasil}
\format{
Um data frame com 5570 municípios e 16 colunas com as seguintes informações adicionais:
\describe{
\item{TIPO}{tipo do item}
\item{CODIGO}{código do item}
\item{DESCRICAO}{descrição do item}
\item{UNIDADE}{unidade de medida (ex: m)}
\item{PRECO_XX}{preço do estado com sigla XX}
}
}
\source{
\url{https://www.caixa.gov.br/poder-publico/modernizacao-gestao/sinapi/}
}
\usage{
municipio
}
\description{
Dataset contendo lista de municípios, códigos ibge e regiões
}
\keyword{datasets}
