% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{necessidade_agua_esgoto}
\alias{necessidade_agua_esgoto}
\title{Calcula as necessidades produção e distribuição}
\usage{
necessidade_agua_esgoto(df)
}
\arguments{
\item{data}{um \code{data.frame} contendo os dados do SNIS}
}
\value{
um \code{data.frame} contendo as densidades
}
\description{
Calcula as necessidades para distribuição de água ($m/hab$) e para produção de água ($m^3/hab$).
}
\examples{
\dontrun{
df <- necessidade_agua_esgoto(df)
}
}
