% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{nome_para_ano}
\alias{nome_para_ano}
\title{Retorna o ano do dataset de população}
\usage{
nome_para_ano(path)
}
\arguments{
\item{path}{Caminho do dataset de população}
}
\value{
ano do dataset de populacao
}
\description{
Retorna o ano do dataset de população
}
\examples{
ano <- nome_para_ano("populaca_estimada_2021")
}
