% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{numero_caminhoes_bau}
\alias{numero_caminhoes_bau}
\title{Número de caminhões baú}
\usage{
numero_caminhoes_bau(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas CO063, CO064, CO065, CO066, CO067 e CO068}
}
\value{
um \code{data.frame} contendo a coluna \code{numero_caminhoes_bau}
}
\description{
Cálcula o número de caminhões baú
}
