% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plano_drenagem}
\alias{plano_drenagem}
\title{PLanos de drenagem municípios brasileiros}
\format{
Uma dataframe com um muncípio por linha e as colunas:
\describe{
\item{codigo_municipio}{Código IBGE do município}
\item{ano_plano}{ano do plano}
\item{investimento}{valor do investimento total previsto para o ano_plano}
}
}
\source{
\url{http://www.snis.gov.br/}
}
\usage{
plano_drenagem
}
\description{
Dataset contendo informaçoes sobre os plano de drenagem
}
\keyword{datasets}
