% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pluviometria}
\alias{pluviometria}
\title{CPRM - dados de pluviometria por município}
\format{
Uma data frame com um muncípio por linhas e as colunas:
\describe{
\item{codigo_municipio}{Código IBGE do município}
\item{municipio}{nome do munícipio}
\item{estado}{estado}
\item{precipitacao_media}{precipitação anual média}
\item{precipitacao_moda}{precipitação anual moda}
}
}
\source{
\url{http://www.snis.gov.br/}
}
\usage{
pluviometria
}
\description{
Dataset contendo informaçoes sobre a precipitação anual por município
}
\keyword{datasets}
