% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{populacao_agrega_municipio}
\alias{populacao_agrega_municipio}
\title{Agrega dados de população por município}
\usage{
populacao_agrega_municipio(tabela)
}
\arguments{
\item{tabela}{é um \code{data.frame} com os dados agregados por setor censitário da população}
}
\value{
um \code{data.frame} com os dados agregados por município
}
\description{
Agrega dados de população por município
}
\examples{
populacao_agrega_municipio(tabela)
}
