% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{precipitacao}
\alias{precipitacao}
\title{Precipitação}
\usage{
precipitacao(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna \code{codigo_municipio}.}
}
\value{
a tabela inicial (\code{data.frame}) com a coluna \code{precipitacao_moda}
}
\description{
Adiciona a coluna de precipitacao_moda na tabela de acordo com o munícipio.
}
\examples{
\dontrun{
tabela <- precipitacao(tabela)
}
}
