% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{preco_unidade_faixa}
\alias{preco_unidade_faixa}
\title{Une tabelas por faixa populacional}
\usage{
preco_unidade_faixa(tabela1, tabela2)
}
\arguments{
\item{tabela1}{o \code{data.frame} que manterá todas as linhas}

\item{tabela2}{o \code{data.frame} com campos adicionais}
}
\value{
um \code{data.frame} com as tabelas unidas
}
\description{
Faz a união entre duas tabelas, a tabela 1 manterá todas suas linhas.
Os valores serão \code{NA} caso não haja uma faixa populacional na tabela2 mas haja na tabela 1.
Ambas tabelas deve ter o campo \code{faixa}.
}
