% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{preenche_atendimento_coleta_indiferenciada}
\alias{preenche_atendimento_coleta_indiferenciada}
\title{Preenche atendimento coleta indiferenciada}
\usage{
preenche_atendimento_coleta_indiferenciada(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} com as colunas "codigo_municipio", "regiao", "faixa", "populacao_urbana_corrente", "atendimento_coleta_indiferenciada_hab"}
}
\value{
um \code{data.frame} contendo a coluna atendimento_relativo_coleta_indiferenciada
}
\description{
\enumerate{
\item Filtra os municipios que tem coleta indiferenciada
\item Calcula o atendimento relativo por regiao e faixa
\item Preenche os municípios que tem coleta indiferenciada e que nao forneceram dados.
Os valores do atendimento é calculado pela populacao_urbana * média relativa de atendimento da regiao e faixa populacional do município.
\item Se a populacao atentida for maior que a populacao urbana define a populacao atendida igual a populacao urbana
}
}
