% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{preenche_caminhoes_bau}
\alias{preenche_caminhoes_bau}
\title{Title}
\usage{
preenche_caminhoes_bau(
  tabela,
  tabela_densidade,
  campo_densidade = "densidade",
  campo_faixa = "faixa"
)
}
\arguments{
\item{campo_faixa}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
