% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{preenche_por_regiao_faixa_populacional}
\alias{preenche_por_regiao_faixa_populacional}
\title{Preenche por regiao e faixa populacional}
\usage{
preenche_por_regiao_faixa_populacional(
  tabela,
  tabela_faixa,
  preencher_campo = "densidade",
  campo_regiao = "regiao",
  campo_faixa = "faixa"
)
}
\arguments{
\item{tabela}{um \code{data.frame} com as colunas a serem preenchidas}

\item{tabela_faixa}{um \code{data.frame} contendo os valores que serão base para o preenchimento devem conter a coluna com os nomes fornecidos em \code{campo} e \code{campo_faixa}}

\item{preencher_campo}{um \code{character} com o nome da coluna a ser preenchida}

\item{campo_regiao}{um \code{character} com o nome da coluna região}

\item{campo_faixa}{um \code{character} com o nome da coluna faixa populacional}
}
\description{
O dado faltante de um estado e uma faixa populacional é preenchido com valores que variam somente com a faixa populacional (\code{tabela_faixa})
}
\examples{
\dontrun{
tabela <- preenche_por_regiao_faixa_populacional(tabela, densidade)
}
}
