% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{preenche_quantidade_coletada}
\alias{preenche_quantidade_coletada}
\title{Preenche quantidade coletada de residuos}
\usage{
preenche_quantidade_coletada(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} com as colunas a serem preenchidas}
}
\value{
o mesmo \code{data.frame} com a coluna preenchida (CO119)
}
\description{
O campo de quantidade coletada de residuos (CO119) é prenchido pela taxa média por regiao e estado.
}
