% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{preenche_situacao}
\alias{preenche_situacao}
\title{Preenche situação
As situações urbana e rural são estimadas mantendo-se a mesma proporção da fonte1.}
\usage{
preenche_situacao(fonte1, fonte2)
}
\arguments{
\item{fonte1}{é a fonte de dados de população mais antiga.}

\item{fonte2}{é a fonte de dados de população mais recente.}
}
\value{
Um \code{data.frame} contendo
}
\description{
Preenche situação
As situações urbana e rural são estimadas mantendo-se a mesma proporção da fonte1.
}
