% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{prepara_regressao}
\alias{prepara_regressao}
\title{Concatena dados para regressão}
\usage{
prepara_regressao(plano, tabela)
}
\arguments{
\item{plano}{um \code{data.frame} contendo as colunas \code{codigo_muncipio} e 'investimento_corrigido'}

\item{tabela}{um \code{data.frame} contendo as colunas \code{codigo_muncipio}, \code{pd} e \code{GE006}}
}
\value{
um \code{data.frame} contendo os campos necessários para regressão linear
}
\description{
Concatena dados para regressão
}
\examples{
\dontrun{
plano <- prepara_regressao(plano, tabela)
}
}
