% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidacao.R
\name{prepara_sankey}
\alias{prepara_sankey}
\title{Prepara tabela para plotar gráfico estilo sankey}
\usage{
prepara_sankey(tabela, colunas)
}
\arguments{
\item{tabela}{o \code{data.frame} longo com dados de investimento}

\item{colunas}{um \code{vector} com a lista de coluns para serem utilizadas.
Os items podem ser "situacao", "destino", "componente", "etapa".}
}
\value{
uma lista (\code{list}) que guarda todos os dados para o sankey plot
}
\description{
Prepara tabela para plotar gráfico estilo sankey
}
