% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis.R
\name{prepara_snis}
\alias{prepara_snis}
\title{Carrega os dados do SNIS}
\usage{
prepara_snis(ano)
}
\arguments{
\item{ano}{ano do SNIS a ser carregado}
}
\value{
um \code{data.frame} contendo os dados do SNIS para o ano especificado
}
\description{
Carrega os dados do SNIS
}
