% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidacao.R
\name{prepare_sankey}
\alias{prepare_sankey}
\title{Prepara tabela para}
\usage{
prepare_sankey(tabela)
}
\arguments{
\item{tabela}{o \code{data.frame} longo com dados de investimento}
}
\value{
uma lista (\code{list}) que guarda todos os dados para o sankey plot
}
\description{
Prepara tabela para
}
