% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{preprocess_censo2022_data}
\alias{preprocess_censo2022_data}
\title{Cria conjunto de dados do censo IBGE 2022}
\usage{
preprocess_censo2022_data()
}
\description{
Exporta dados do censo 2022 para CSV e ser utilizado como base de calculo
}
