% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{preprocess_sinisa_data}
\alias{preprocess_sinisa_data}
\title{SINISA Pre-processing
Reads the downloaded SINISA files and preprocess them into csv files}
\usage{
preprocess_sinisa_data(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\description{
SINISA Pre-processing
Reads the downloaded SINISA files and preprocess them into csv files
}
