% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R
\name{processa_sinapi_v2025}
\alias{processa_sinapi_v2025}
\title{SINAPI Formato 2025
Ingestao de Insumos e Composições Não Desonerados (Sem Desoneração)}
\usage{
processa_sinapi_v2025(ano, mes)
}
\arguments{
\item{ano}{um \code{number} com o ano desejado}

\item{mes}{um \code{number} com o mês desejado}
}
\description{
SINAPI Formato 2025
Ingestao de Insumos e Composições Não Desonerados (Sem Desoneração)
}
