% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{projecao_reaproveitamento}
\alias{projecao_reaproveitamento}
\title{Projeção do reaproveitamento}
\usage{
projecao_reaproveitamento(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas "estado", "faixa", "reaproveitamento_relativo", "total_residuos_projecao"}
}
\value{
um \code{data.frame} contendo a coluna reaproveitamento_relativo
}
\description{
Projeção do reaproveitamento
}
