% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projeto_coleta_esgoto}
\alias{projeto_coleta_esgoto}
\title{Projeto tipo para coleta de esgoto}
\format{
Um data frame com 54 items serviços e materiais e 9 colunas, sendo estas:
\describe{
\item{codigo}{Código do item no SINAPI}
\item{grupo}{Grupo de custos (ex. serviço ou material)}
\item{tipo}{Tipo de custo (ex. demolição, excavação)}
\item{07a}{quantidade de material para cenário população pequena e de alto custo}
\item{07b}{quantidade de material para cenário população pequena e de baixo custo}
\item{08a}{quantidade de material para cenário população média e de alto custo}
\item{08b}{quantidade de material para cenário população média e de baixo custo}
\item{09a}{quantidade de material para cenário população grande e de alto custo}
\item{09b}{quantidade de material para cenário população grande e de baixo custo}
}
}
\usage{
projeto_coleta_esgoto
}
\description{
Dataset contendo lista de serviços e materiais para criar infraestrutura de coleta de esgoto.
}
\keyword{datasets}
