% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projeto_predominancia_tipo_producao}
\alias{projeto_predominancia_tipo_producao}
\title{Projeto tipo para distribuicao de água}
\format{
Um data frame com 27 estados e 3 colunas, sendo estas:
\describe{
\item{estado}{Sigal da unidade federativa}
\item{subterranea}{Predominância em \% de extração de água superficial}
\item{superficial}{Predominância em \% de extração de água subterrânea}
}
}
\source{
\url{}
}
\usage{
projeto_predominancia_tipo_producao
}
\description{
Dataset contendo lista de serviços e materiais para criar infraestrutura de distribuicao de água.
}
\keyword{datasets}
