% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{quantidade_caminhoes_basculantes}
\alias{quantidade_caminhoes_basculantes}
\title{Número de caminhões baú}
\usage{
quantidade_caminhoes_basculantes(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas CO063, CO064, CO065, CO066, CO067 e CO068}
}
\value{
um \code{data.frame} contendo a coluna \code{quantidade_caminhoes_basculantes}
}
\description{
Cálcula o número de caminhões baú
}
