% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{quantidade_compostagem_municipio}
\alias{quantidade_compostagem_municipio}
\title{Quantidade de resíduos destinada à compostagem (t/ano)}
\usage{
quantidade_compostagem_municipio(tabela)
}
\arguments{
\item{tabela}{contendo a coluna UP003}
}
\value{
tabela contendo a quantidade de resíduos destinados a compostagem por município
}
\description{
Agrupa por município os dados do SNIS (prestadores)
e estimar a quantidade de resíduos destinados a compostagem.
}
\examples{
\dontrun{tabela <- quantidade_compostagem_municipio(tabela)}
}
