% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{regionaliza100}
\alias{regionaliza100}
\title{Regionaliza 100\% da demanda}
\usage{
regionaliza100(tabela, campo)
}
\arguments{
\item{tabela}{é um \code{data.frame} contendo as colunas \code{demanda_triagem}, \code{estado} e \code{faixa} populacional}

\item{campo}{é um \code{character} com o nome da coluna de demanda}
}
\value{
um \code{data.frame} onde a \verb{demanda_triagem esta regionalizada}
}
\description{
Regionaliza 100\% da demanda
}
\examples{
\dontrun{
tabela <- regionalizacao_triagem(tabela, potencial)
}
}
