% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{regionaliza_compostagem}
\alias{regionaliza_compostagem}
\title{Regionaliza demanda por compostagem}
\usage{
regionaliza_compostagem(tabela, cenario)

regionaliza_compostagem(tabela, cenario)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo numero_municipio}

\item{cenario}{um \code{character} contendo o nome do cenario ("A" "B" ou "C")}
}
\value{
um \code{data.frame} contendo a coluna \code{demanda_tranbordo}

um \code{data.frame} contendo a coluna \code{demanda_tranbordo}
}
\description{
Faz a regionalização da demanda em acordo com o cenário de regionalização.

Faz a regionalização da demanda em acordo com o cenário de regionalização.
}
\examples{
\dontrun{

}
\dontrun{

}
}
