% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{regionaliza_faixa5e6}
\alias{regionaliza_faixa5e6}
\title{Regionaliza faixa populacional 5 e 6}
\usage{
regionaliza_faixa5e6(tabela, potencial, campo_demanda, campo_potencial)
}
\arguments{
\item{tabela}{é um \code{data.frame} contendo as colunas \code{demanda_triagem}, \code{regiao}, \code{estado} e \code{faixa} populacional}

\item{potencial}{é um \code{data.frame}contendo as colunas \code{regiao} e \code{potencial_regionalização}}

\item{campo}{é um \code{character} com o nome da coluna de demanda}
}
\value{
um \code{data.frame} onde a coluna de demanda esta regionalizada
}
\description{
Regionaliza faixa populacional 5 e 6
}
\examples{
\dontrun{
tabela <- regionaliza_faixa5e6(tabela, potencial)
}
}
