% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{regionaliza_transbordo}
\alias{regionaliza_transbordo}
\title{Regionaliza demanda por transbordo}
\usage{
regionaliza_transbordo(tabela, cenario)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo numero_municipio}

\item{cenario}{um \code{character} contendo o nome do cenario ("A" "B" ou "C")}
}
\value{
um \code{data.frame} contendo a coluna \code{demanda_tranbordo}
}
\description{
A demanda por transbordo é considerada 1 estacao de transbordo e 1 caminhao por municipio
}
\examples{
\dontrun{

}
}
