% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{regressao_drenagem}
\alias{regressao_drenagem}
\title{Regressão Investimento em Drenagem}
\usage{
regressao_drenagem(plano)
}
\arguments{
\item{plano}{é um \code{data.frame} contendo os valores do plano}
}
\value{
objeto da classe \code{lm} contendo os parâmetros da regressão
}
\description{
Calcula a regressão linear entre densidade de investimento em drenagem e o índice PD
}
\examples{
\dontrun{
equacao <- regressao_drenagem(plano)
}
}
