% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drenagem.R
\name{remove_nao_criticos}
\alias{remove_nao_criticos}
\title{Filtra municípios críticos}
\usage{
remove_nao_criticos(tabela)
}
\arguments{
\item{tabela}{contendo a coluna código do município}
}
\value{
a tabela de entrada mantendo apenas os municípios críticos
}
\description{
Filtra municípios críticos
}
