% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_shiny_classes}
\alias{remove_shiny_classes}
\title{Remove items com classes shiny}
\usage{
remove_shiny_classes(lista)
}
\arguments{
\item{lista}{}
}
\value{
list sem items que são classes shiny
}
\description{
Percorre todos os itens procurando itens em que a classe contenha shiny no nome e as remove.
}
\examples{
a <- list()
remove_shiny_classes(a)
}
