% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{residuos_cobertura_preprocess}
\alias{residuos_cobertura_preprocess}
\title{Residuos Pre-processing Cobertura
Read the SINISA waste coverage data and process it into a data frame.}
\usage{
residuos_cobertura_preprocess(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\value{
A data frame containing the processed waste coverage date
}
\description{
Residuos Pre-processing Cobertura
Read the SINISA waste coverage data and process it into a data frame.
}
