% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{residuos_destinacao_preprocess}
\alias{residuos_destinacao_preprocess}
\title{Residuos Pre-processing Destinação Final
Read the SINISA waste final destination data and process it into a data frame.}
\usage{
residuos_destinacao_preprocess(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\value{
A data frame containing the processed waste final destination data
}
\description{
Residuos Pre-processing Destinação Final
Read the SINISA waste final destination data and process it into a data frame.
}
