% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{residuos_manejo_preprocess}
\alias{residuos_manejo_preprocess}
\title{Residuos Pre-processing Manejo
Read the SINISA waste management data and process it into a data frame.}
\usage{
residuos_manejo_preprocess(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\value{
A data frame containing the processed waste management data
}
\description{
Residuos Pre-processing Manejo
Read the SINISA waste management data and process it into a data frame.
}
