% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{residuos_preprocess}
\alias{residuos_preprocess}
\title{Residuos Pre-processing
Read the SINISA waste data and process it into a data frame.}
\usage{
residuos_preprocess(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\value{
A data frame containing the processed waste data
}
\description{
Residuos Pre-processing
Read the SINISA waste data and process it into a data frame.
}
