% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinisa.R
\name{residuos_unidades_preprocess}
\alias{residuos_unidades_preprocess}
\title{Residuos Pre-processing Unidades de Processamento
Read the SINISA waste processing units data and process it into a data frame.}
\usage{
residuos_unidades_preprocess(year)
}
\arguments{
\item{year}{Year of the data to process (e.g., 2022)}
}
\value{
A data frame containing the processed waste processing units data
}
\description{
Residuos Pre-processing Unidades de Processamento
Read the SINISA waste processing units data and process it into a data frame.
}
