% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{restaura_parametros_json}
\alias{restaura_parametros_json}
\title{Restaura os parâmetros da ultima sessão calculada (JSON)}
\usage{
restaura_parametros_json()
}
\value{
um boleano indicando se os parâmetros foram restaurados com sucesso
}
\description{
Restaura os parâmetros da ultima sessão calculada (JSON)
}
