% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{restaura_parametros_padrao}
\alias{restaura_parametros_padrao}
\title{Restaura os parâmertos padrão}
\usage{
restaura_parametros_padrao()
}
\value{
um boleano indicando se os parâmetros foram restaurados com sucesso
}
\description{
Restaura os parâmertos padrão
}
