% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{rodar_modulo_demografico}
\alias{rodar_modulo_demografico}
\title{Módulo demográfico}
\usage{
rodar_modulo_demografico(input, projecao, tema)
}
\arguments{
\item{input}{estrutura de dados (\code{reactive}) que guarda os parâmetros da interface gráfica}

\item{projecao}{é o \code{data.frame} com o estado atual da projeção populacional}

\item{tema}{define qual tema será caculado (\code{agua} ou \code{esgoto})}
}
\value{
um \code{data.frame} contendo as necessidades de intraestrutura para um dado \code{tema}
}
\description{
Calcula as necessidades
}
\examples{
\dontrun{
tabela <- rodar_modulo_demografico(input, projecao, "agua")
}
}
