% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{rodar_modulo_financeiro}
\alias{rodar_modulo_financeiro}
\title{Title}
\usage{
rodar_modulo_financeiro(input, orcamentario)
}
\arguments{
\item{orcamentario}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
