% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{rodar_modulo_financeiro_agua}
\alias{rodar_modulo_financeiro_agua}
\title{Title}
\usage{
rodar_modulo_financeiro_agua(input, orcamentario, tipo = "agua")
}
\arguments{
\item{orcamentario}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
