% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{rodar_modulo_orcamentario}
\alias{rodar_modulo_orcamentario}
\title{Title}
\usage{
rodar_modulo_orcamentario(input, demografico)
}
\arguments{
\item{demografico}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
