% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{rodar_modulo_orcamentario_agua}
\alias{rodar_modulo_orcamentario_agua}
\title{Módulo orçamentário para água}
\usage{
rodar_modulo_orcamentario_agua(input, demografico)
}
\arguments{
\item{input}{estrutura de dados (\code{reactive}) que guarda os parâmetros da interface gráfica}

\item{demografico}{estrutura de dados (\code{data.frame}) que armazena os resultados do módulo demográfico}
}
\value{
estrutura de dados (\code{data.frame}) que armazenas os resultados do módulo orçamentário
}
\description{
Módulo orçamentário para água
}
\examples{
\dontrun{
orca <- rodar_modulo_orcamentario_agua(input, demografico)
}
}
