% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{rodar_modulo_rural_agua}
\alias{rodar_modulo_rural_agua}
\title{Módulo de cálculo para demanda rural de água}
\usage{
rodar_modulo_rural_agua(state)
}
\arguments{
\item{input}{estrutura de dados (\code{reactive}) que guarda os parâmetros da interface gráfica}

\item{taxas_projecao}{um \code{data.frame} contendo as taxas de projecao populacional}
}
\value{
um \code{data.frame} contendo as necessidade de investimentos e todos campos utilizados
}
\description{
Esta função organiza a ordem de execução das tarefas necessárias
para o cálculo de necessidades de investimento em sistema de abastecimento de água.
}
