% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{rodar_projecao_populacional}
\alias{rodar_projecao_populacional}
\title{Calcula a projeção populacional}
\usage{
rodar_projecao_populacional(state)
}
\arguments{
\item{input}{estrutura de dados (\code{reactive}) que guarda os parâmetros da interface gráfica}
}
\value{
estado da aplicação (\code{list}) que guarda todos os resultado e parâmetros dos cálculos
}
\description{
Calcula a projeção populacional
}
\examples{
\dontrun{
app_state <- rodar_projecao_populacional(app_state)
}
}
