% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{salva_base_calculo}
\alias{salva_base_calculo}
\title{Salva a base de c치lculo da Necessidade de Investimento em Saneamento}
\usage{
salva_base_calculo(df, componente, fonte, ano)
}
\arguments{
\item{df}{Data frame com os dados a serem salvos}

\item{componente}{Componente da base de c치lculo}

\item{fonte}{Fonte dos dados}

\item{ano}{Ano da base de c치lculo}
}
\value{
Nenhum valor retornado
}
\description{
Salva a base de c치lculo da Necessidade de Investimento em Saneamento
}
