% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{salva_dados_base_calculo}
\alias{salva_dados_base_calculo}
\title{Salva todos os arquivos de dados/base_calculo em um único arquivo .rda em data/}
\usage{
salva_dados_base_calculo()
}
\description{
Esta função lê todos os arquivos CSV do diretório 'dados/base_calculo/' e salva um arquivo RDA chamado 'base_calculo.rda' no diretório 'data/'. Cada arquivo será um data.frame em uma lista nomeada pelo nome do arquivo (sem extensão).
}
