% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametros.R
\name{salva_parametros}
\alias{salva_parametros}
\title{Salva os parametros da interface gráfica no estado da aplicação}
\usage{
salva_parametros(state, input, name)
}
\arguments{
\item{state}{estrutura de dados (\code{list}) que guarda o estado atual da aplicação}

\item{input}{estrutura de dados (\code{reactive}) que guarda os parâmetros da interface gráfica}

\item{name}{um \code{character} como nome do campo para ser guardado o dados do parâmetro \code{input}}
}
\value{
a estrutura de dados (\code{list}) do novo estado atual da aplicação
}
\description{
Salva os parametros da interface gráfica no estado da aplicação
}
\examples{
\dontrun{
salva_parametros_projecao(state, input, "teste")
}
}
