% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametros.R
\name{salva_parametros_agua_esgoto}
\alias{salva_parametros_agua_esgoto}
\title{Salva os parâmetros do cálculo de investimento em água e esgoto}
\usage{
salva_parametros_agua_esgoto(state, input)
}
\arguments{
\item{state}{estrutura de dados (\code{list}) que guarda o estado atual da aplicação}

\item{input}{estrutura de dados (\code{reactive}) que guarda os parâmetros da interface gráfica}
}
\value{
a estrutura de dados (\code{list}) do novo estado atual da aplicação
}
\description{
Salva os parâmetros do cálculo de investimento em água e esgoto
}
\examples{
\dontrun{
salva_parametros_agua_esgoto(state, input)
}
}
