% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametros.R
\name{salva_parametros_projecao}
\alias{salva_parametros_projecao}
\title{Salva os parametros do cálculo da projeção populacional}
\usage{
salva_parametros_projecao(state, input)
}
\arguments{
\item{state}{estrutura de dados (\code{list}) que guarda o estado atual da aplicação}

\item{input}{estrutura de dados (\code{reactive}) que guarda os parâmetros da interface gráfica}
}
\value{
a estrutura de dados (\code{list}) do novo estado atual da aplicação
}
\description{
Salva os parametros do cálculo da projeção populacional
}
\examples{
\dontrun{
salva_parametros_projecao(state, input)
}
}
