% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{salva_resultado_intermediario}
\alias{salva_resultado_intermediario}
\title{Salva resultados intermedi치rios do calculo da necessidade de investimento}
\usage{
salva_resultado_intermediario(df, nome, cenario = "base")
}
\arguments{
\item{df}{Data frame com os dados a serem salvos}

\item{nome}{Nome do arquivo a ser salvo}

\item{cenario}{Cen치rio de c치lculo}
}
\value{
Nenhum valor retornado
}
\description{
Salva resultados intermedi치rios do calculo da necessidade de investimento
}
