% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_calculo.R
\name{save_base_calculo_data}
\alias{save_base_calculo_data}
\title{Salva todos os arquivos de dados/base_calculo em um único arquivo .rda em data/}
\usage{
save_base_calculo_data()
}
\description{
Esta função lê todos os arquivos CSV do diretório 'dados/base_calculo/' e salva um arquivo RDA chamado 'base_calculo.rda' no diretório 'data/'. Cada arquivo será um data.frame em uma lista nomeada pelo nome do arquivo (sem extensão).
}
