% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_state.R
\name{save_state}
\alias{save_state}
\title{Salva o estado da aplicação}
\usage{
save_state(app_state)
}
\arguments{
\item{app_state}{estrutura de dados (\code{list}) que guarda o estado atual da aplicação}
}
\description{
Salva o estado da aplicação
}
\examples{
\dontrun{
save_state(app_state)
}
}
