% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sinapi}
\alias{sinapi}
\title{SINAPI Dezembro de 2021}
\format{
Um data frame com 12414 items e 31 variáveis
\describe{
\item{TIPO}{tipo do item}
\item{CODIGO}{código do item}
\item{DESCRICAO}{descrição do item}
\item{UNIDADE}{unidade de medida (ex: m)}
\item{PRECO_XX}{27 colunas com preço do estado (sigla XX) ex: PRECO_AC é o preço no ACRE.}
}
}
\source{
\url{https://www.caixa.gov.br/poder-publico/modernizacao-gestao/sinapi/}
}
\usage{
sinapi
}
\description{
Dataset contendo preços e índices para a construção civil.
}
\keyword{datasets}
