% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R
\name{sinapi_id_to_name}
\alias{sinapi_id_to_name}
\title{Transforma id sinapi em nome legível}
\usage{
sinapi_id_to_name(id)
}
\value{
um \code{character} com o nome
}
\description{
Transforma id sinapi em nome legível
}
