% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snis2021}
\alias{snis2021}
\title{SNIS 2021}
\format{
A data frame with 5570 rows and 608 variables
\describe{
\item{codigo_municipio}{codigo ibge do municipio}
\item{\code{CODIGO}}{codigo do indicador ou informacao do SNIS ex: AG001}
... todas variáveis do snis
}
}
\source{
\url{http://www.snis.gov.br/}
}
\usage{
snis2021
}
\description{
Dataset contendo as informações e indicadores do SNIS do ano de 2021.
}
\keyword{datasets}
