% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R, R/esgoto.R
\docType{data}
\name{snis_fields}
\alias{snis_fields}
\title{Necessidade de Investimento - Componente Água - Situação Urbana}
\format{
An object of class \code{character} of length 12.

An object of class \code{character} of length 12.
}
\usage{
snis_fields

snis_fields
}
\description{
Necessidade de Investimento - Componente Água - Situação Urbana

Necessidade de Investimento - Componente Esgoto - Situação Urbana
}
\keyword{datasets}
