% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis.R
\name{snis_prepara_agua}
\alias{snis_prepara_agua}
\title{Prepara os dados do SNIS para o componente água}
\usage{
snis_prepara_agua(df)
}
\arguments{
\item{df}{Data frame com os dados do SNIS}
}
\value{
Data frame com os dados do SNIS para o componente água
}
\description{
Prepara os dados do SNIS para o componente água
}
