% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snis_rs}
\alias{snis_rs}
\title{SNIS - Resíduos sólidos dados por prestador}
\format{
Uma lista de data frame indexados por ano de referência.
\describe{
\item{Código}{Código IBGE do município}
\item{Nome}{nome do munícipio}
\item{UF}{estado}
\item{Ano}{ano de referência}
\item{UndCod}{código do prestador}
\item{YYXXXX}{Indicadores e informações}
}
}
\source{
\url{http://www.snis.gov.br/}
}
\usage{
snis_rs
}
\description{
Dataset contendo informaçoes das unidades de resíduos sólidos por prestador
}
\keyword{datasets}
