% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{soma_por_estado_faixa}
\alias{soma_por_estado_faixa}
\title{Soma todas as variáveis numéricas agrupando por estado e faixa populacional}
\usage{
soma_por_estado_faixa(tabela, campo_estado = "Estado", campo_faixa = "faixa")
}
\arguments{
\item{tabela}{tabela contendo as colunas de população, estado}

\item{campo_estado}{parâmetro opcional para definir o campo que contém a sigla do estado (default: "Estado")}

\item{campo_faixa}{parâmetro opcional para definir o campo que contém a faixa populaciona (default: "faixa")}
}
\value{
tabela contendo todas variáveis somadas por grupos de estado e faixa populacional
}
\description{
Soma todas as variáveis numéricas agrupando por estado e faixa populacional
}
\examples{
Estado <- c(rep("AC",4), rep("AL",4))
faixa <- c(rep(1,2), rep(3,2), rep(2,2), rep(4,2))
valor <- c(1, 1, 2, 2, 3, 3 , 4, 4)
input <- dplyr::tibble(Estado, faixa, valor)
output <- soma_por_estado_faixa(input)
}
