% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{soma_por_faixa}
\alias{soma_por_faixa}
\title{Soma todas as variáveis numéricas agrupando por faixa populacional}
\usage{
soma_por_faixa(tabela, campo_faixa = "faixa")
}
\arguments{
\item{tabela}{tabela contendo as colunas de população, estado}

\item{campo_faixa}{parâmetro opcional para definir o campo que contém a faixa populaciona (default: "faixa")}
}
\value{
tabela contendo todas variáveis somadas por faixa populacional
}
\description{
Soma todas as variáveis numéricas agrupando por faixa populacional
}
\examples{
faixa <- c(rep(1, 2), rep(3, 2), rep(2, 2), rep(4, 2))
valor <- c(1, 1, 2, 2, 3, 3, 4, 4)
input <- dplyr::tibble(faixa, valor)
output <- soma_por_faixa(input)
}
