% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localidades.R
\name{states_acronym}
\alias{states_acronym}
\title{Lista de siglas do estados mais distrito federal}
\usage{
states_acronym()
}
\value{
lista com siglas dos estados em ordem alfabética
}
\description{
Lista de siglas do estados mais distrito federal
}
\examples{
states <- states_acronym()
}
