% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{substitui_por_faixa_populacional}
\alias{substitui_por_faixa_populacional}
\title{Substitui por faixa populacional}
\usage{
substitui_por_faixa_populacional(
  tabela,
  tabela_faixa,
  preencher_campo = "densidade",
  campo_faixa = "faixa"
)
}
\arguments{
\item{tabela}{um \code{data.frame} com as colunas a serem preenchidas}

\item{tabela_faixa}{um \code{data.frame} contendo os valores que serão base para o preenchimento devem conter a coluna com os nomes fornecidos em \code{campo} e \code{campo_faixa}}

\item{preencher_campo}{um \code{character} com o nome da coluna a ser preenchida}

\item{campo_faixa}{um \code{character} com o nome da coluna da faixa populacional}
}
\value{
o mesmo \code{data.frame} com o campo substituido
}
\description{
O dados faltante de um campo é preenchido com valores que variam somente com a faixa populacional (\code{tabela_faixa})
}
\examples{
\dontrun{
tabela <- substitui_por_faixa_populacional(tabela, densidade)
}
}
