% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{tabela_preco_unidade_residuos}
\alias{tabela_preco_unidade_residuos}
\title{Title}
\usage{
tabela_preco_unidade_residuos(input, name)
}
\arguments{
\item{name}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
