% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igp.R
\name{taxa_correcao_ipea}
\alias{taxa_correcao_ipea}
\title{Retorna a taxa de correção de preços}
\usage{
taxa_correcao_ipea(ipea_code)
}
\arguments{
\item{ipea_code}{Código do índice no IPEA Data.}
}
\value{
Um \code{data.frame} contendo o índice IGP-DI mais atualizado, com as colunas:
\itemize{
\item{data}{Data do índice.}
\item{igp}{Valor do índice.}
}
}
\description{
Retorna o índice IGP-DI mês a mês até o dia mais atual.
Fonte: http://www.ipeadata.gov.br/
}
\examples{
igp <- taxa_correcao_ipea("IGP12_INCC12")
}
