% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{taxa_geracao_residuos}
\alias{taxa_geracao_residuos}
\title{Taxa de geração de resíduos (t/hab)}
\usage{
taxa_geracao_residuos(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas residuo_coletado_ton_ano, atendimento_coleta_indiferenciada_hab, residuo_coletado_ton_ano_litoral e atendimento_coleta_indiferenciada_hab_litoral}
}
\value{
um \code{data.frame} contendo as colunas taxa_geracao_residuos, taxa_geracao_residuos_litoral
}
\description{
Calcula geração de resíduos para municípios com pesagem
}
