% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{tbl_longa_deficit_agua_rural}
\alias{tbl_longa_deficit_agua_rural}
\title{Cria tabela longa de deficit do componente água e situacao rural}
\usage{
tbl_longa_deficit_agua_rural(tabela)
}
\arguments{
\item{tabela}{contendo as colunas:
\itemize{
\item{estado}
\item{regiao}
\item{deficit_urbana}
}}
}
\value{
tabela contendo os campos:
\itemize{
\item{estado}
\item{regiao}
\item{componente}
\item{situacao}
\item{etapa}
\item{deficit}
}
}
\description{
Cria tabela longa de deficit do componente água e situacao rural
}
