% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto.R
\name{tbl_longa_deficit_esgoto}
\alias{tbl_longa_deficit_esgoto}
\title{Cria tabela longa de deficit do componente esgoto}
\usage{
tbl_longa_deficit_esgoto(tabela)
}
\arguments{
\item{tabela}{contendo as colunas:
\itemize{
\item{estado}
\item{regiao}
\item{deficit_urbana}
}}
}
\value{
tabela contendo os campos:
\itemize{
\item{estado}
\item{regiao}
\item{componente}
\item{situacao}
\item{subsistema}
\item{deficit}
}
}
\description{
Cria tabela longa de deficit do componente esgoto
}
