% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{tbl_longa_deficit_esgoto_rural}
\alias{tbl_longa_deficit_esgoto_rural}
\title{Cria tabela longa de deficit do componente esgoto e situacao rural}
\usage{
tbl_longa_deficit_esgoto_rural(tabela)
}
\arguments{
\item{tabela}{contendo as colunas:
\itemize{
\item{estado}
\item{regiao}
\item{deficit_rural}
}}
}
\value{
tabela contendo os campos:
\itemize{
\item{estado}
\item{regiao}
\item{componente}
\item{situacao}
\item{subsistema}
\item{deficit}
}
}
\description{
Cria tabela longa de deficit do componente esgoto e situacao rural
}
