% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{tbl_longa_deficit_residuos}
\alias{tbl_longa_deficit_residuos}
\title{Cria tabela longa de deficit do componente residuos}
\usage{
tbl_longa_deficit_residuos(tabela)
}
\arguments{
\item{tabela}{contendo as colunas:
\itemize{
\item{estado}
\item{regiao}
\item{deficit_coleta_seletiva}
\item{deficit_coleta_indiferenciada}
}}
}
\value{
tabela contendo os campos:
\itemize{
\item{estado}
\item{regiao}
\item{componente}
\item{situacao}
\item{etapa}
\item{deficit}
}
}
\description{
Cria tabela longa de deficit do componente residuos
}
