% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{tbl_longa_investimento_residuos}
\alias{tbl_longa_investimento_residuos}
\title{Cria tabela longa de necessidade de investimento do componente residuos}
\usage{
tbl_longa_investimento_residuos(tabela)
}
\arguments{
\item{tabela}{contendo as colunas:
\itemize{
\item\link{investimento_reposicao_aterro}
\item\link{investimento_reposicao_compostagem}
\item\link{investimento_reposicao_triagem}
\item\link{investimento_reposicao_coleta_indiferenciada}
\item\link{investimento_reposicao_coleta_seletiva}
\item\link{investimento_expansao_aterro}
\item\link{investimento_expansao_compostagem}
\item\link{investimento_expansao_triagem}
\item\link{investimento_expansao_transbordo}
\item\link{investimento_expansao_coleta_indiferenciada}
\item\link{investimento_expansao_coleta_seletiva}
}}
}
\value{
tabela contendo os campos:
\itemize{
\item{estado}
\item{regiao}
\item{componente}
\item{situacao}
\item{destino}
\item{subsistema}
}
}
\description{
Cria tabela longa de necessidade de investimento do componente residuos
}
