% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{tbl_longa_investimentos_agua}
\alias{tbl_longa_investimentos_agua}
\title{Cria tabela longa de necessidade de investimento do componente água}
\usage{
tbl_longa_investimentos_agua(tabela)
}
\arguments{
\item{tabela}{contendo as colunas:
\itemize{
\item{deficit_urbana}
\item{custo_expansao_producao_agua}
\item{custo_reposicao_producao_agua}
\item{custo_reposicao_distribuicao_agua}
}}
}
\value{
tabela contendo os campos:
\itemize{
\item{estado}
\item{regiao}
\item{componente}
\item{situacao}
\item{destino}
\item{subsistema}
\item{necessidade_investimento}
}
}
\description{
Cria tabela longa de necessidade de investimento do componente água
}
