% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{tbl_longa_investimentos_agua_rural}
\alias{tbl_longa_investimentos_agua_rural}
\title{Cria tabela longa de necessidade de investimento do componente água}
\usage{
tbl_longa_investimentos_agua_rural(tabela)
}
\arguments{
\item{tabela}{contendo as colunas:
\itemize{
\item{custo_expansao_distribuicao_agua}
\item{custo_expansao_producao_agua}
\item{custo_reposicao_producao_agua}
\item{custo_reposicao_distribuicao_agua}
}}
}
\value{
tabela contendo os campos:
\itemize{
\item{estado}
\item{regiao}
\item{componente}
\item{situacao}
\item{destino}
\item{etapa}
}
}
\description{
Cria tabela longa de necessidade de investimento do componente água
}
