% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto.R
\name{tbl_longa_investimentos_esgoto}
\alias{tbl_longa_investimentos_esgoto}
\title{Cria tabela longa de necessidade de investimento do componente esgoto}
\usage{
tbl_longa_investimentos_esgoto(tabela)
}
\arguments{
\item{tabela}{contendo as colunas:
\itemize{
\item{custo_expansao_coleta_esgoto}
\item{custo_expansao_tratamento_esgoto}
\item{custo_reposicao_tratamento_esgoto}
\item{custo_reposicao_coleta_esgoto}
}}
}
\value{
tabela contendo os campos:
\itemize{
\item{estado}
\item{regiao}
\item{componente}
\item{situacao}
\item{destino}
\item{subsistema}
}
}
\description{
Cria tabela longa de necessidade de investimento do componente esgoto
}
