% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{to_power_bi}
\alias{to_power_bi}
\title{Exporta os dados para Power BI}
\usage{
to_power_bi(tabela, file)
}
\arguments{
\item{tabela}{tabela de dados com as necessidades de investimentos com o seguintes campos:
"estado";"regiao";"destino";"subsistema";"necessidade_investimento";"componente";"situacao";"pais"}

\item{file}{caminho do arquivo de saída}
}
\value{
Não retorna nada, apenas salva o arquivo
}
\description{
Exporta os dados para Power BI
}
