% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{update_ibge_censo}
\alias{update_ibge_censo}
\title{Atualiza dados do censo IBGE}
\usage{
update_ibge_censo(ano)
}
\arguments{
\item{year}{é um \code{number} com o ano desejado}
}
\value{
um \code{logical} dizendo se a atualização ocorreu com sucesso
}
\description{
Atualiza dados do censo IBGE
}
