% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R
\name{update_sinapi}
\alias{update_sinapi}
\title{Atualiza dados do SINAPI}
\usage{
update_sinapi(ano, mes)
}
\arguments{
\item{year}{é um \code{number} com o ano desejado}

\item{month}{é um \code{number} com o mês desejado}
}
\value{
um \code{logical} dizendo se a atualização ocorreu com sucesso
}
\description{
Atualiza dados do SINAPI
}
