% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis_ap.R
\name{update_snis_ap}
\alias{update_snis_ap}
\title{Atualiza dados do SNIS-AP}
\usage{
update_snis_ap(ano)
}
\arguments{
\item{ano}{é um \code{number} com o ano desejado}
}
\value{
um \code{logical} dizendo se a atualização ocorreu com sucesso
}
\description{
Atualiza dados do SNIS-AP
}
