% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_snis_rs.R
\name{update_snis_rs}
\alias{update_snis_rs}
\title{Atualiza dados do SNIS-RS}
\usage{
update_snis_rs(ano)
}
\arguments{
\item{ano}{é um \code{number} com o ano desejado}
}
\value{
um \code{logical} dizendo se a atualização ocorreu com sucesso
}
\description{
Atualiza dados do SNIS-RS
}
