% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{valida_base_calculo}
\alias{valida_base_calculo}
\title{Verifica a integridade da base de cálculo}
\usage{
valida_base_calculo(df, componente)
}
\arguments{
\item{df}{Data frame a ser validado}

\item{componente}{Componente da base de cálculo}
}
\value{
\code{TRUE} se a base de cálculo for válida, \code{FALSE} caso contrário
}
\description{
Verifica a integridade da base de cálculo
}
