% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{workspace_dir}
\alias{workspace_dir}
\title{Retorna o diret처rio de trabalho onde os dados ser찾o armazenados}
\usage{
workspace_dir()
}
\value{
The directory where datasets are stored
}
\description{
Retorna o diret처rio de trabalho onde os dados ser찾o armazenados
}
\examples{
wd <- workspace_dir()
}
