% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{adiciona_preco_unidade_residuos}
\alias{adiciona_preco_unidade_residuos}
\title{Preço unidade de resíduos}
\usage{
adiciona_preco_unidade_residuos(input, name, precos)
}
\arguments{
\item{input}{estrutura de dados (\code{reactive}) que guarda os parâmetros da interface gráfica}

\item{name}{um \code{character} com o nome do projeto}

\item{precos}{uma \code{list} com os preços por faixa}
}
\value{
os parâmetros da interface gráfica com os preços corrigidos
}
\description{
Salva preços por faixa populacional no formato da interface gráfica.
}
