% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecao.R
\name{adicionar_proporcao_urbana_rural}
\alias{adicionar_proporcao_urbana_rural}
\title{Proporção relativa para populações urbana e rural.}
\usage{
adicionar_proporcao_urbana_rural(tabela)
}
\arguments{
\item{tabela}{é um tibble contendo os dados de populações total, urbana e rural.}
}
\value{
Um tibble() contendo as colunas relativa_urbana e relativa_rural.
}
\description{
Calcula frações da população urbana e rural em relação a população total.
}
\examples{
\dontrun{
tabela <- adicionar_proporcao_urbana_rural(tabela)
}
}
