% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{atendimento_relativo_coleta_seletiva}
\alias{atendimento_relativo_coleta_seletiva}
\title{Atendimento relativo para coleta seletiva}
\usage{
atendimento_relativo_coleta_seletiva(tabela)
}
\arguments{
\item{tabela}{}
}
\description{
Atendimento relativo para coleta seletiva
}
\examples{
\dontrun{

}
}
