% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{atendimento_relativo_residuos}
\alias{atendimento_relativo_residuos}
\title{Atendimento relativo para residuo}
\usage{
atendimento_relativo_residuos(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} com as colunas \code{atendimento_coleta_indiferenciada_hab}, \code{CO050}, \code{populacao_total_corrente} e \code{populacao_urbana_corrente}.}
}
\value{
um \code{data.frame} contendo as coliunas \code{atendimento_relativo_total}, \code{atendimento_relativo_urbano} e \code{atendimento_relativo_rural}
}
\description{
Calcula o atendimento relativo total, urbano e rural.
}
