% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{capacidade_instalada_coleta_seletiva}
\alias{capacidade_instalada_coleta_seletiva}
\title{Capacidade instalada coleta seletiva}
\usage{
capacidade_instalada_coleta_seletiva(tabela, valor)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{POP_URB}, \code{atendimento_relativo_seletiva_urbano} e \code{densidade_caminhoes_bau }}

\item{valor}{valor um \code{double} contendo o preço de um caminhão bau}
}
\value{
um \code{data.frame} contendo a coluna capacidade_instalada_coleta_seletiva
}
\description{
Capacidade instalada coleta seletiva
}
