% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{capacidade_instalada_triagem}
\alias{capacidade_instalada_triagem}
\title{Capacidade instalada de triagem}
\usage{
capacidade_instalada_triagem(tabela, vida_util)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{meta_reaproveitamento}, \code{total_residuos} e \code{preco_unidade_triagem}}

\item{vida_util}{um \code{number} com o numero de anos da vida útil do sistema de triagem}
}
\value{
um \code{data.frame} com a coluna \code{capacidade_instalada_triagem}
}
\description{
Capacidade instalada de triagem
}
