% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_zhelper.R
\name{check_dataset_integrity}
\alias{check_dataset_integrity}
\title{Verifica a integridade do conjunto de dados}
\usage{
check_dataset_integrity(name)
}
\arguments{
\item{name}{é um \code{character} com o nome do conjunto de dados}
}
\description{
Verifica a integridade do conjunto de dados
}
