% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{deficit_coleta_indiferenciada}
\alias{deficit_coleta_indiferenciada}
\title{Title}
\usage{
deficit_coleta_indiferenciada(tabela)
}
\arguments{
\item{tabela}{}
}
\description{
Title
}
\examples{
\dontrun{

}
}
