% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{demanda_compostagem}
\alias{demanda_compostagem}
\title{Demanda por compostagem}
\usage{
demanda_compostagem(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{meta_compostagem}, \code{total_residuos_projecao} e \code{quantidade_compostagem}.}
}
\value{
um \code{data.frame} com a coluna \code{demanda_compostagem}
}
\description{
Demanda por compostagem
}
