% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{domicilios_adequados_com_esgoto}
\alias{domicilios_adequados_com_esgoto}
\title{Número de domicílios com demanda adequada de esgoto}
\usage{
domicilios_adequados_com_esgoto(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{domicilios_projecao} e \code{deficit_esgoto_relativo_rural}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{domicilios_deficit_esgoto}.
}
\description{
Calcula o número de domicílios com déficit de sistemas de estação de esgoto.
}
