% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{domicilios_com_deficit_agua}
\alias{domicilios_com_deficit_agua}
\title{Número de domícilios com deficit de água}
\usage{
domicilios_com_deficit_agua(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas \code{V001}, \code{V001_projecao} e \code{deficit_agua_relativo_rural}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{domicilios_deficit_agua}.
}
\description{
Calcula o número de domícilios com déficit de sistemas de abastecimento de água.
}
