% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{download_censo_raw}
\alias{download_censo_raw}
\title{Cria conjunto de dados do censo IBGE}
\usage{
download_censo_raw(year)
}
\arguments{
\item{year}{ano que deseja baixa`1}
}
\value{
um \code{data.frame} com a estimativa populacional pelo Censo
}
\description{
Cria conjunto de dados do censo IBGE
}
\examples{
download_censo_raw()
}
