% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua_rural.R
\name{fazer_projecao_domicilio}
\alias{fazer_projecao_domicilio}
\title{Faz projeção de domícilios}
\usage{
fazer_projecao_domicilio(
  tabela,
  ano_inicial,
  ano_final,
  atualizar_moradores_por_domicilio = FALSE
)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo a coluna \code{codigo_municipio}}

\item{ano_inicial}{um \code{number} contendo o ano do conjunto de dados de população (V002)}

\item{ano_final}{um \code{number} contendo o ano da projeção}

\item{atualizar_moradores_por_domicilio}{um \code{boolean} dizendo se existe uma atualização da relação domicílio/habitantes (V003). O padrão é \code{FALSE}.}
}
\value{
um \code{data.frame} contendo a coluna adicional \code{taxa_crescimento}.
}
\description{
Projeta o número de domícilio pela projeção populacional e considera que a relação domicílio/habitantes se mantém constante.
}
