% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{get_censo_labels}
\alias{get_censo_labels}
\title{Retorna os rótulos dos dados do IBGE disponíveis (somente censo)}
\usage{
get_censo_labels()
}
\value{
um vetor com os dados disponiveis
}
\description{
Retorna os rótulos dos dados do IBGE disponíveis (somente censo)
}
\examples{
\dontrun{
df <- get_sinapi_data(sinapi)
}
}
