% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{get_censo_labels}
\alias{get_censo_labels}
\title{Retorna os rótulos dos arquivos de censo disponíveis na base de cálculo}
\usage{
get_censo_labels()
}
\value{
Um vetor com os nomes dos arquivos de censo disponíveis (sem extensão).
}
\description{
Esta função lista os arquivos na pasta 'dados/base_calculo' que contêm o padrão "censo" no nome.
}
