% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R
\name{get_sinapi_labels}
\alias{get_sinapi_labels}
\title{Retorna os rótulos dos dados do SINAPI disponívei}
\usage{
get_sinapi_labels()
}
\value{
um vetor com os dados disponiveis
}
\description{
Retorna os rótulos dos dados do SINAPI disponívei
}
\examples{
\dontrun{
df <- get_sinapi_data(sinapi)
}
}
