% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_universan.R
\name{get_sinapi_labels}
\alias{get_sinapi_labels}
\title{Retorna lista de dados sinapi disponíveis}
\usage{
get_sinapi_labels()
}
\value{
Uma lista nomeada com os nomes dos arquivos sinapi disponíveis.
}
\description{
Esta função retorna uma lista nomeada com os arquivos disponíveis na pasta 'dados/base_calculo' que
começam com o prefixo "sinapi_". Os nomes dos arquivos são retornados sem o caminho e sem a extensão.
}
