% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sinapi.R, R/data_universan.R
\name{get_sinapi_labels}
\alias{get_sinapi_labels}
\title{Retorna os rótulos dos dados do SINAPI disponívei}
\usage{
get_sinapi_labels()

get_sinapi_labels()
}
\value{
um vetor com os dados disponiveis

Uma lista nomeada com os nomes dos arquivos sinapi disponíveis.
}
\description{
Esta função retorna uma lista nomeada com os arquivos disponíveis na pasta 'dados/base_calculo' que
começam com o prefixo "sinapi_". Os nomes dos arquivos são retornados sem o caminho e sem a extensão.
}
\examples{
\dontrun{
df <- get_sinapi_data(sinapi)
}
}
