% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{get_snis_data}
\alias{get_snis_data}
\title{Ferramentas para Cálculo das Necessidades de Investimento
para o Saneamento Água e Esgoto
Retorna dados do SNIS}
\usage{
get_snis_data(snis, fields)
}
\arguments{
\item{snis}{caminho do SNIS}

\item{fields}{lista com campos a serem retornados}
}
\value{
um tibble() contendo os dados do SNIS
}
\description{
Ferramentas para Cálculo das Necessidades de Investimento
para o Saneamento Água e Esgoto
Retorna dados do SNIS
}
\examples{
\dontrun{
df <- get_snis_data("snis_2020", c("AG001", "AG010"))
}
}
