% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igp.R
\name{get_taxa_igp}
\alias{get_taxa_igp}
\title{Calculo da taxa de correção de preço - IGP}
\usage{
get_taxa_igp(data_inicial, data_final)
}
\arguments{
\item{data_inicial}{é um \code{date} contendo a data em que o preço foi atribuido}

\item{data_final}{é um \code{date} contendo uma data mais recente pra qual se quer corrigir o preço}

\item{igp}{é um \code{data.frame} com a série histórica do IGP com as colunas data e igp(taxa).}
}
\value{
um \code{double} que é a taxa pela qual o preço deve ser multiplicada para corrígi-lo.
}
\description{
A função retorna a taxa de correção de preço para um determinado intervalo de datas.
O algoritmo busca as taxas mais próximas das datas passadas e retorna a taxa de correção.
}
\examples{
data(igp)
valor <- function(igp, as.Date("2016-01-01"), as.Date("2021-01-01"))
}
