% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{ignora_residuo_litoraneo}
\alias{ignora_residuo_litoraneo}
\title{Ignora quantidade de residuo litorâneo}
\usage{
ignora_residuo_litoraneo(tabela)

ignora_residuo_litoraneo(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas CO119, CO164 e litoral}
}
\value{
um \code{data.frame} onde os municipios litoraneos não tem dados de quantidade de residuos

um \code{data.frame} onde os municipios litoraneos não tem dados de quantidade de residuos
}
\description{
Ignora quantidade de residuo litorâneo

Ignora quantidade de residuo litorâneo
}
\examples{
\dontrun{
tbl <- ignora_residuo_litoraneo(tbl)
}
\dontrun{
tbl <- ignora_residuo_litoraneo(tbl)
}
}
