% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{investimento_compostagem}
\alias{investimento_compostagem}
\title{Investimento em compostagem}
\usage{
investimento_compostagem(tabela, vida_util)
}
\arguments{
\item{vida_util}{}
}
\description{
Investimento em compostagem
}
\examples{
\dontrun{

}
}
