% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{investimento_drenagem}
\alias{investimento_drenagem}
\title{Cálculo da Necessidade de Investimento para Drenagem Urbana}
\usage{
investimento_drenagem(state)
}
\arguments{
\item{input}{estrutura de dados (\code{reactive}) que guarda os parâmetros da interface gráfica}
}
\value{
estado da aplicação (\code{list}) que guarda todos os resultado e parâmetros dos cálculos
}
\description{
Cálculo da Necessidade de Investimento para Drenagem Urbana
}
\examples{
\dontrun{
state <- investimento_drenagem(state)
}
}
