% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{investimento_residuos}
\alias{investimento_residuos}
\title{Title}
\usage{
investimento_residuos(state)
}
\arguments{
\item{input}{}
}
\description{
Title
}
\examples{
\dontrun{
app_state <- investimento_residuos(app_state)
}
}
