% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{investimento_residuos}
\alias{investimento_residuos}
\title{Investimento em residuos}
\usage{
investimento_residuos(state)
}
\arguments{
\item{state}{estrutura de dados (\code{list}) que guarda o estado atual da aplicação}
}
\value{
estado da aplicação (\code{list}) que guarda todos os resultado e parâmetros dos cálculos
}
\description{
Investimento em residuos
}
\examples{
\dontrun{
app_state <- investimento_residuos(app_state)
}
}
