% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{investimento_residuos_total}
\alias{investimento_residuos_total}
\title{Totaliza os valores de investimentos em resíduos}
\usage{
investimento_residuos_total(tabela)
}
\arguments{
\item{tabela}{contendo os valores de reposição e investimento em resíduos:
\itemize{
\item\link{resposicao_aterro}
\item\link{resposicao_compostagem}
\item\link{resposicao_triagem}
\item\link{resposicao_coleta_regular}
\item\link{resposicao_coleta_seletiva}
\item\link{investimento_aterro}
\item\link{investimento_compostagem}
\item\link{investimento_triagem}
\item\link{investimento_coleta_regular}
\item\link{investimento_coleta_seletiva}
}}
}
\value{
um \code{data.frame} contendo as colunas adicionais \code{investimento_reposicao}, \code{investimento_expansao} e \code{investimento_total}
}
\description{
Consolida os valores de investimentos para reposicao, expansão e total
}
\examples{
\dontrun{

}
}
