% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esgoto_rural.R
\name{investimento_rural_esgoto}
\alias{investimento_rural_esgoto}
\title{Necessidade de investimento rural para sistema de estação de esgoto}
\usage{
investimento_rural_esgoto(tabela)
}
\arguments{
\item{tabela}{um \code{data.frame} contendo as colunas:
\code{custo_esgoto_individual}, \code{custo_esgoto_producao_rural}, \code{custo_esgoto_distribuicao_rural},
\code{fracao_investimento_individual_esgoto}, \code{fracao_investimento_coletivo_esgoto} e
\code{domicilios_deficit_esgoto}}
}
\value{
um \code{data.frame} contendo as colunas adicionais:
investimento_expansao_individual, investimento_expansao_tratamento_esgoto e investimento_expansao_coleta_esgoto
}
\description{
Calcula a necessidade de investimento para expansão do sistema de estação de esgoto.
}
