% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{investimento_triagem}
\alias{investimento_triagem}
\title{Investimento em triagem}
\usage{
investimento_triagem(tabela, vida_util)
}
\arguments{
\item{vida_util}{}
}
\description{
Investimento em triagem
}
\examples{
\dontrun{

}
}
