% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ibge_populacao.R
\name{list_files_from_ftp}
\alias{list_files_from_ftp}
\title{List of files/folders in a FTP folder}
\usage{
list_files_from_ftp(url)
}
\arguments{
\item{url}{}
}
\value{
list of files and folders
}
\description{
List of files/folders in a FTP folder
}
\examples{
list_files_from_ftp("ftp.ibge.gov.br/Censos/")
}
