% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos_modulo.R
\name{meta_plansab_residuo}
\alias{meta_plansab_residuo}
\title{Adiciona metas de atendimento por região}
\usage{
meta_plansab_residuo(tabela)
}
\arguments{
\item{tabela}{}
}
\description{
Adiciona metas de atendimento por região
}
\examples{
\dontrun{

}
}
