% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipio}
\alias{municipio}
\title{Municípios do Brasil}
\format{
Um data frame com 5570 municípios e as seguintes 6 colunas:
\describe{
\item{codigo_municipio}{código IBGE do munícipio}
\item{municipio}{Nome do município}
\item{regiao}{Região do Brasil}
\item{regiao_sigla}{Sigla da região do Brasil}
\item{estado}{Unidade federativa}
\item{estado_sigla}{Sigla da unidade federativa}
}
}
\source{
\url{https://servicodados.ibge.gov.br/api/v1/localidades}
}
\usage{
municipio
}
\description{
Dataset contendo classificação de localização dos municípios (região e UF).
}
\keyword{datasets}
