% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{preenche_densidade_caminhao_bau}
\alias{preenche_densidade_caminhao_bau}
\title{Preenche caminhao basculante}
\usage{
preenche_densidade_caminhao_bau(tabela_por_municipio, estado_faixa)
}
\arguments{
\item{tabela_por_municipio}{um \code{data.frame} com as colunas "codigo_municipio", "faixa", "tem_coleta_seletiva", "atendimento_coleta_seletiva_hab", "quantidade_caminhoes_basculantes"}

\item{estado_faixa}{um \code{data.frame} com as colunas "estado", "faixa", "populacao_urbana_corrente", "atendimento_coleta_seletiva_hab", "tem_coleta_seletiva"}
}
\value{
um \code{data.frame} contendo a coluna atendimento_relativo_coleta_seletiva
}
\description{
\enumerate{
\item Filtra os municipios que tem coleta seletiva
\item Calcula a quantidade de caminhões basculante por faixa
}
}
