% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.R
\name{rodar_modelo}
\alias{rodar_modelo}
\title{Roda todos os modelos de cálculo de investimento}
\usage{
rodar_modelo(state)
}
\arguments{
\item{state}{}
}
\description{
Roda todos os modelos de cálculo de investimento
}
\examples{
\dontrun{
rodar_modelo(state)
}
}
