% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aguaesgoto.R
\name{rodar_modulo_demografico}
\alias{rodar_modulo_demografico}
\title{Title}
\usage{
rodar_modulo_demografico(input, projecao)
}
\arguments{
\item{projecao}{}
}
\description{
Title
}
\examples{
\dontrun{
rodar_modulo_demografico
}
}
