% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agua.R
\name{rodar_modulo_demografico}
\alias{rodar_modulo_demografico}
\title{Módulo demográfico}
\usage{
rodar_modulo_demografico(input, projecao, tema)
}
\arguments{
\item{tema}{define qual tema será caculado (\code{agua} ou \code{esgoto})}
}
\description{
Calcula as necessidades
}
\examples{
\dontrun{
rodar_modulo_demografico
}
}
